% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acgme.R
\docType{data}
\name{acgme}
\alias{acgme}
\title{ACGME OBGYN Residency Data}
\format{
A data frame with the following columns:
\describe{
\item{Program directors}{The names of program directors for OBGYN residency programs.}
\item{name}{The names of OBGYN residency programs.}
}
}
\source{
Data was obtained from the ACGME website: \url{https://apps.acgme.org/ads/Public/Programs/Search}
}
\usage{
acgme
}
\value{
A tibble listing ACGME-accredited OBGYN residency programs and
their program directors.
}
\description{
This dataset provides information about Obstetricians and Gynecologists (OBGYN) residency programs accredited by the Accreditation Council for Graduate Medical Education (ACGME).
}
\examples{
# Load the ACGME OBGYN Residency Data
data(acgme)

# Inspect the dataset
print(acgme)

# Get a summary of the dataset
summary(acgme)

# Perform data analysis and exploration

}
\seealso{
Other datasets: 
\code{\link{ACOG_Districts}},
\code{\link{cityStateToLatLong}},
\code{\link{fips}},
\code{\link{taxonomy}}
}
\concept{datasets}
\keyword{dataset}
