% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_and_save_tint.R
\name{apply_and_save_tint}
\alias{apply_and_save_tint}
\title{Apply and Save Image Tint}
\usage{
apply_and_save_tint(
  image_path,
  alpha = 0.5,
  palette_name = "viridis",
  save_dir = ".",
  n_colors = 5
)
}
\arguments{
\item{image_path}{Character string specifying the file path of the image to be tinted.}

\item{alpha}{Numeric value between 0 and 1 specifying the degree of the tint.
Higher values will result in stronger tints.}

\item{palette_name}{Character string specifying the color palette to use for the tint.
Default is 'viridis'.}

\item{save_dir}{Character string specifying the directory to save the tinted images.}

\item{n_colors}{Integer specifying the number of different colors to use from the color palette.}
}
\value{
Saves tinted images in the specified directory.
}
\description{
This function applies a tint to an image and saves the tinted image.
The tints are determined from a given color palette.
}
