% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal_tables_write.R
\name{arsenal_tables_write2word}
\alias{arsenal_tables_write2word}
\title{Writes an Arsenal table object to a Word document.}
\usage{
arsenal_tables_write2word(object, filename, output_dir = NULL)
}
\arguments{
\item{object}{An object to be written to Word, typically an Arsenal table.}

\item{filename}{The filename (without extension) for the Word document.}

\item{output_dir}{Directory where the Word document should be written.
Defaults to a session-specific folder inside \code{\link[=tempdir]{tempdir()}}.}
}
\value{
Called for its side effect of saving a Word document to \code{word_path}. Returns \code{NULL} invisibly.
}
\description{
Writes an Arsenal table object to a Word document.
}
\examples{
\dontrun{
arsenal_tables_write2word(my_table, "output_table")
}
}
\seealso{
Other table: 
\code{\link{max_table}()},
\code{\link{min_table}()},
\code{\link{table_calculate_percentages}()},
\code{\link{table_calculate_proportion}()},
\code{\link{table_generate_overall}()}
}
\concept{table}
