% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{calculate_proportion_variable}
\alias{calculate_proportion_variable}
\title{Calculate the Proportion Variable}
\usage{
calculate_proportion_variable(tabyl_result)
}
\arguments{
\item{tabyl_result}{The tabulation result data frame.}
}
\value{
The proportion variable as a percentage.
}
\description{
This function calculates and returns the proportion variable, which is the maximum percentage in the tabulation result.
}
