% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_block_group_overlap_map.R
\name{create_block_group_overlap_map}
\alias{create_block_group_overlap_map}
\title{Function to create and export a map showing block group overlap with isochrones}
\usage{
create_block_group_overlap_map(
  bg_data,
  isochrones_data,
  output_dir = "figures/"
)
}
\arguments{
\item{bg_data}{A SpatialPolygonsDataFrame representing block group data.}

\item{isochrones_data}{A SpatialPolygonsDataFrame representing isochrone data.}

\item{output_dir}{Directory path for exporting the map files. Default is "figures/".}
}
\value{
None
}
\description{
This function creates a map that displays block groups and their overlap with isochrones.
}
\examples{
\dontrun{
# Create and export the map with the default output directory
create_block_group_overlap_map(block_groups, isochrones_joined_map)

# Create and export the map with a custom output directory
create_block_group_overlap_map(block_groups, isochrones_joined_map, "custom_output/")
}

}
