% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_formula_maker.R
\name{create_formula}
\alias{create_formula}
\title{Create a Formula for Poisson Model}
\usage{
create_formula(data, response_var, random_effect = NULL)
}
\arguments{
\item{data}{A dataframe containing the predictor and response variables.}

\item{response_var}{The name of the response variable in the dataframe.}

\item{random_effect}{Optional. The name of the random effect variable for the formula.}
}
\value{
A formula object suitable for modeling in R.
}
\description{
This function creates a formula for a Poisson model based on the provided data,
response variable, and optional random effect.
}
\examples{
# Example usage:
response_variable <- "days"
random_effect_term <- "name"  # Change this to the desired random effect variable
df3_filtered <- data.frame(days = c(5, 10, 15), age = c(30, 40, 50), name = c("A", "B", "C"))
formula <- create_formula(df3_filtered, response_variable, random_effect_term)
formula

}
