% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\docType{data}
\name{fips}
\alias{fips}
\title{Data of FIPS codes}
\format{
A tibble with 243 rows and 10 variables:
\describe{
\item{state}{Two-letter postal abbreviation.}
\item{state_name}{Full state name.}
\item{state_fips}{Two-digit state FIPS code.}
\item{county_fips}{Three-digit county FIPS code.}
\item{fips}{Combined five-digit state and county code.}
\item{class}{Geography class indicator.}
\item{county}{County name.}
\item{county_ansi}{County ANSI code.}
\item{county_short}{Simplified county name.}
\item{state_ansi}{State ANSI code.}
}
}
\source{
\url{https://github.com/kjhealy/fips-codes/blob/master/state_and_county_fips_master.csv}
}
\usage{
fips
}
\value{
A tibble containing Federal Information Processing Standards
(FIPS) codes for states and counties.
}
\description{
3 columns with 51 observations.
}
\seealso{
Other datasets: 
\code{\link{ACOG_Districts}},
\code{\link{acgme}},
\code{\link{cityStateToLatLong}},
\code{\link{taxonomy}}
}
\concept{datasets}
\keyword{datasets}
