% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_leaflet_base_map.R
\name{generate_leaflet_base_map}
\alias{generate_leaflet_base_map}
\title{Generate Leaflet Base Map}
\usage{
generate_leaflet_base_map()
}
\value{
A Leaflet map object.
}
\description{
This function creates a Leaflet BASE map with specific configurations, including the base tile layer, scale bar,
default view settings, and layers control.
}
\examples{
map <- generate_leaflet_base_map()
}
