% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\name{get_census_data}
\alias{get_census_data}
\title{Get Census data for all states' block groups}
\usage{
get_census_data(us_fips_list)
}
\arguments{
\item{us_fips_list}{A vector of state FIPS codes.  For example Colorado == 08.}

\item{us_fips}{A vector of state FIPS codes for which Census data is to be retrieved.}

\item{vintage}{The vintage year for Census data (default is 2019).}
}
\value{
A dataframe containing Census data for all states' block groups.

A dataframe containing Census data for all states' block groups.
}
\description{
This function retrieves Census data for all states' block groups by looping
over the specified list of state FIPS codes.
}
\examples{
us_fips_list <- tigris::fips_codes \%>\%
dplyr::select(state_code, state_name) \%>\%
dplyr::distinct(state_code, .keep_all = TRUE) \%>\%
dplyr::filter(state_code < 56) \%>\%
dplyr::select(state_code) \%>\%
dplyr::pull()

us_fips_list <- c("01", "02", "04", "05", "06", "08", "09", "10", "11", "12", "13", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "44", "45", "46", "47", "48", "49", "50", "51", "53", "54", "55")

all_census_data <- get_census_data(us_fips_list)

Get Census data for all states' block groups

This function retrieves Census data for all states' block groups by looping
over the specified list of state FIPS codes.

}
