% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{get_most_common}
\alias{get_most_common}
\title{Get the Most Common Value}
\usage{
get_most_common(tabyl_result, variable_name)
}
\arguments{
\item{tabyl_result}{The tabulation result data frame.}

\item{variable_name}{The name of the categorical variable.}
}
\value{
The most common value.
}
\description{
This function calculates and returns the most common value from the tabulation result.
}
