% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycomb_map_by_acog_districts.R
\name{honeycomb_generate_maps}
\alias{honeycomb_generate_maps}
\title{Generate Hexagon Maps by ACOG District}
\usage{
honeycomb_generate_maps(
  physician_sf,
  acog_districts_sf,
  trait_map = "all",
  honey_map = "all",
  grid_size = c(0.3, 0.3),
  specific_district = NULL
)
}
\arguments{
\item{physician_sf}{A sf object containing physician data with coordinates.}

\item{acog_districts_sf}{A sf object containing the grouped ACOG districts.}

\item{trait_map}{A string specifying the trait map (default is "all").}

\item{honey_map}{A string specifying the honey map (default is "all").}

\item{grid_size}{A numeric vector of length 2 specifying the grid size for the hexagon map (default is c(0.3, 0.3)).}

\item{specific_district}{A string or NULL to specify a specific district for generating the map (default is NULL, which processes all districts).}
}
\value{
A ggplot object of the generated map for the specified or all districts.
}
\description{
This function generates hexagon maps for ACOG districts.
}
