% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_load_libraries.R
\name{install_and_load_libraries}
\alias{install_and_load_libraries}
\title{Install and Load Required Libraries}
\usage{
install_and_load_libraries()
}
\value{
A character vector containing the names of the loaded packages.
}
\description{
This function installs and loads the necessary libraries for your analysis.
It checks if the packages are already installed and installs them if needed.
It then loads the installed packages.
}
\examples{
install_and_load_libraries()

}
