% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_exploratory_packages.R
\name{install_exploratory_packages}
\alias{install_exploratory_packages}
\title{Install Exploratory Packages}
\usage{
install_exploratory_packages()
}
\value{
None
}
\description{
This function checks if required packages for Exploratory Data Analysis are installed and installs them if necessary.
It includes packages from both GitHub and CRAN repositories.
}
\details{
This function checks and installs necessary packages for Exploratory Data Analysis in R.

This function checks and installs the following packages:
\itemize{
\item From GitHub: anonymizer, glue, exploratory_func.
\item From CRAN: backports, psych, qlcMatrix, quanteda, textshape, mnormt, sparsesvd, docopt, proxyC.
}
}
\examples{
\dontrun{
# Call the function to install and load packages
install_exploratory_packages()
}
}
\seealso{
\code{\link{install.packages}} to install packages from CRAN.
\code{\link{install_github}} to install packages from GitHub.
}
