% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_normalizer.R
\name{is_po_box}
\alias{is_po_box}
\title{Detect PO Box Addresses}
\usage{
is_po_box(x)
}
\arguments{
\item{x}{Character vector of address strings to check.}
}
\value{
Logical vector indicating whether each address is a PO Box.
}
\description{
Checks if an address string represents a Post Office Box. Detects common
variations including "PO BOX", "P O BOX", and "POST OFFICE BOX".
}
\examples{
is_po_box("PO Box 12345")
is_po_box("123 Main Street")

}
\seealso{
Other address-normalization: 
\code{\link{ascii_norm}()},
\code{\link{caps}()},
\code{\link{has_street_number}()},
\code{\link{normalize_address_df}()},
\code{\link{normalize_directionals}()},
\code{\link{normalize_state}()},
\code{\link{normalize_suffix}()},
\code{\link{normalize_units}()},
\code{\link{normalize_zip5}()},
\code{\link{strip_suite}()}
}
\concept{address-normalization}
