% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_matching_data.R
\name{load_matching_data}
\alias{load_matching_data}
\title{Load Matching Data for a Residency Program}
\usage{
load_matching_data(location, year)
}
\arguments{
\item{location}{The location of the residency program.}

\item{year}{The year of the matching data.}
}
\value{
A processed data frame containing the matching data, or NULL in case of an error.
}
\description{
This function reads and processes matching data for a given residency program and year.
}
