% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create_leaflet_base.R
\name{map_create_leaflet_base}
\alias{map_create_leaflet_base}
\title{Create a Leaflet Base Map}
\usage{
map_create_leaflet_base()
}
\value{
A Leaflet map object.
}
\description{
This function creates a Leaflet BASE map with specific configurations, including the base tile layer, scale bar,
default view settings, and layers control.
}
\examples{
map <- map_create_leaflet_base()
}
