% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanci.R
\name{meanci}
\alias{meanci}
\title{meanci}
\usage{
meanci(dataset, group_by_vars, ...)
}
\arguments{
\item{dataset}{The dataset to use}

\item{group_by_vars}{The grouping variable(s) to use, given as a character vector}

\item{...}{The variable(s) to calculate means and CIs for}
}
\description{
Calculate n, mean and upper/lower confidence intervals for one or more variables,
while grouping by one or more variables
}
\examples{

# Calculate mean and CI for `disp`, grouping by `cyl`, using mtcars data
mtcars \%>\% meanci(group_by_vars = c('cyl'), hp)

# Calculate mean and CI for `hp`, grouping by `cyl` and `gear`, using mtcars data
mtcars \%>\% meanci(group_by_vars = c('cyl', 'gear'), hp)

}
