% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_indicators.R
\docType{data}
\name{MEDICAL_SCHOOL_INDICATORS}
\alias{MEDICAL_SCHOOL_INDICATORS}
\title{Medical School Affiliation Patterns}
\format{
Character vector of medical school text patterns.
}
\usage{
MEDICAL_SCHOOL_INDICATORS
}
\description{
String patterns indicating medical school affiliation. Confidence: 0.97.
}
\examples{
length(MEDICAL_SCHOOL_INDICATORS)

}
\seealso{
\code{\link{classify_academic_affiliation}}
}
\keyword{datasets}
