% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mysterycall-deprecated}
\alias{mysterycall-deprecated}
\alias{arsenal_tables_write2word}
\alias{calculate_intersection_overlap_and_save}
\alias{check_normality}
\alias{clean_phase_1_results}
\alias{clean_phase_2_data}
\alias{create_and_plot_interaction}
\alias{create_density_plot}
\alias{create_formula}
\alias{create_individual_isochrone_plots}
\alias{create_isochrones}
\alias{create_isochrones_for_dataframe}
\alias{create_line_plot}
\alias{create_scatter_plot}
\alias{download_large_file}
\alias{format_pct}
\alias{genderize_physicians}
\alias{geocode_unique_addresses}
\alias{get_census_data}
\alias{hrr}
\alias{hrr_generate_maps}
\alias{map_create_acog_districts_sf}
\alias{map_create_base}
\alias{map_create_block_group_overlap}
\alias{map_create_leaflet_base}
\alias{map_create_physician_dot}
\alias{max_table}
\alias{min_table}
\alias{most_common_gender_training_academic}
\alias{physician_age}
\alias{plot_and_save_emmeans}
\alias{plot_census_age_distribution}
\alias{remove_constant_vars}
\alias{remove_near_zero_var}
\alias{rename_columns_by_substring}
\alias{retrieve_clinician_data}
\alias{run_mystery_caller_workflow}
\alias{run_mystery_caller_workflow_with_logging}
\alias{save_quality_check_table}
\alias{search_and_process_npi}
\alias{search_by_taxonomy}
\alias{split_and_save}
\alias{states_where_physicians_were_NOT_contacted}
\alias{summarize_census_data}
\alias{table_calculate_percentages}
\alias{table_calculate_proportion}
\alias{table_generate_overall}
\alias{table_write_pdf}
\alias{validate_and_remove_invalid_npi}
\alias{search_npi}
\alias{test_and_process_isochrones}
\alias{process_and_save_isochrones}
\alias{progress_tracker}
\alias{progress_tracker_start}
\alias{progress_tracker_finish}
\alias{progress_tracker_fail}
\alias{progress_tracker_update}
\alias{progress_tracker_summary}
\title{Deprecated functions in mysterycall}
\usage{
arsenal_tables_write2word(...)

calculate_intersection_overlap_and_save(...)

check_normality(...)

clean_phase_1_results(...)

clean_phase_2_data(...)

create_and_plot_interaction(...)

create_density_plot(...)

create_formula(...)

create_individual_isochrone_plots(...)

create_isochrones(...)

create_isochrones_for_dataframe(...)

create_line_plot(...)

create_scatter_plot(...)

download_large_file(...)

format_pct(...)

genderize_physicians(...)

geocode_unique_addresses(...)

get_census_data(...)

hrr(...)

hrr_generate_maps(...)

map_create_acog_districts_sf(...)

map_create_base(...)

map_create_block_group_overlap(...)

map_create_leaflet_base(...)

map_create_physician_dot(...)

max_table(...)

min_table(...)

most_common_gender_training_academic(...)

physician_age(...)

plot_and_save_emmeans(...)

plot_census_age_distribution(...)

remove_constant_vars(...)

remove_near_zero_var(...)

rename_columns_by_substring(...)

retrieve_clinician_data(...)

run_mystery_caller_workflow(...)

run_mystery_caller_workflow_with_logging(...)

save_quality_check_table(...)

search_and_process_npi(...)

search_by_taxonomy(...)

split_and_save(...)

states_where_physicians_were_NOT_contacted(...)

summarize_census_data(...)

table_calculate_percentages(...)

table_calculate_proportion(...)

table_generate_overall(...)

table_write_pdf(...)

validate_and_remove_invalid_npi(...)

search_npi(input_data, ...)

test_and_process_isochrones(input_file, ...)

process_and_save_isochrones(input_file, chunk_size = 25, ...)

progress_tracker(...)

progress_tracker_start(...)

progress_tracker_finish(...)

progress_tracker_fail(...)

progress_tracker_update(...)

progress_tracker_summary(...)
}
\arguments{
\item{input_data}{A data frame or CSV path containing \code{first} and \code{last} columns.}

\item{chunk_size}{Deprecated.}
}
\description{
These names were used before \code{mysterycall} adopted a consistent \code{mysterycall_} prefix.
They emit deprecation warnings and forward all arguments to the current
equivalent.
}
\keyword{internal}
