% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_normalizer.R
\name{mysterycall_is_po_box}
\alias{mysterycall_is_po_box}
\title{Detect PO Box Addresses}
\usage{
mysterycall_is_po_box(x)
}
\arguments{
\item{x}{Character vector of address strings to check.}
}
\value{
Logical vector indicating whether each address is a PO Box.
}
\description{
Checks if an address string represents a Post Office Box. Detects common
variations including "PO BOX", "P O BOX", and "POST OFFICE BOX".
}
\examples{
mysterycall_is_po_box("PO Box 12345")
mysterycall_is_po_box("123 Main Street")

}
\seealso{
Other address-normalization: 
\code{\link{mysterycall_ascii_norm}()},
\code{\link{mysterycall_caps}()},
\code{\link{mysterycall_has_street_number}()},
\code{\link{mysterycall_normalize_directionals}()},
\code{\link{mysterycall_normalize_state}()},
\code{\link{mysterycall_normalize_suffix}()},
\code{\link{mysterycall_normalize_units}()},
\code{\link{mysterycall_normalize_zip5}()},
\code{\link{mysterycall_strip_suite}()},
\code{\link{normalize_address_df}()}
}
\concept{address-normalization}
