% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{mysterycall_log_cache_hit}
\alias{mysterycall_log_cache_hit}
\title{Log cache hit}
\usage{
mysterycall_log_cache_hit(what, n_items)
}
\arguments{
\item{what}{What was loaded from cache}

\item{n_items}{Number of cached items}
}
\value{
\code{invisible(NULL)}.
}
\description{
Log cache hit
}
\seealso{
Other logging utilities: 
\code{\link{mysterycall_progress_tracker}()},
\code{\link{mysterycall_progress_fail}()},
\code{\link{mysterycall_progress_finish}()},
\code{\link{mysterycall_progress_start}()},
\code{\link{mysterycall_progress_summary}()},
\code{\link{mysterycall_progress_update}()},
\code{\link{mysterycall_log_error}()},
\code{\link{mysterycall_log_info}()},
\code{\link{mysterycall_log_progress}()},
\code{\link{mysterycall_log_save}()},
\code{\link{mysterycall_log_step}()},
\code{\link{mysterycall_log_step_complete}()},
\code{\link{mysterycall_log_success}()},
\code{\link{mysterycall_log_warning}()},
\code{\link{mysterycall_workflow_start}()}
}
\concept{logging utilities}
