% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{mysterycall_log_to_file}
\alias{mysterycall_log_to_file}
\title{Write to log file if configured}
\usage{
mysterycall_log_to_file(msg)
}
\arguments{
\item{msg}{Message to write}
}
\description{
Write to log file if configured
}
\keyword{internal}
