% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create_block_group_overlap.R
\name{mysterycall_map_block_group}
\alias{mysterycall_map_block_group}
\title{Function to create and export a map showing block group overlap with isochrones}
\usage{
mysterycall_map_block_group(bg_data, isochrones_data, output_dir = "figures/")
}
\arguments{
\item{bg_data}{A SpatialPolygonsDataFrame representing block group data.}

\item{isochrones_data}{A SpatialPolygonsDataFrame representing isochrone data.}

\item{output_dir}{Directory path for exporting the map files. Default is "figures/".}
}
\value{
Called for its side effects of saving the block group overlap map as HTML and PNG inside \code{output_dir}. Returns \code{NULL} invisibly.
}
\description{
This function creates a map that displays block groups and their overlap with isochrones.
The map is exported as an HTML file and a PNG image.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Create and export the map with the default output directory
mysterycall_map_block_group(block_groups, isochrones_joined_map)

# Create and export the map with a custom output directory
mysterycall_map_block_group(block_groups, isochrones_joined_map, "custom_output/")

\dontshow{\}) # examplesIf}
}
\seealso{
Other mapping: 
\code{\link{mysterycall_create_isochrones}()},
\code{\link{mysterycall_isochrones_for_df}()},
\code{\link{mysterycall_map_acceptance_rate}()},
\code{\link{mysterycall_map_acog_districts}()},
\code{\link{mysterycall_map_base}()},
\code{\link{mysterycall_map_leaflet}()},
\code{\link{mysterycall_map_physicians}()},
\code{\link{mysterycall_plot_density}()},
\code{\link{mysterycall_plot_isochrones}()},
\code{\link{mysterycall_plot_line}()},
\code{\link{mysterycall_plot_scatter}()}
}
\concept{mapping}
