% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create_leaflet_base.R
\name{mysterycall_map_leaflet}
\alias{mysterycall_map_leaflet}
\title{Create a Leaflet Base Map}
\usage{
mysterycall_map_leaflet()
}
\value{
Invisibly returns the Leaflet map object.
}
\description{
This function creates a Leaflet BASE map with specific configurations, including the base tile layer, scale bar,
default view settings, and layers control.
}
\examples{
\donttest{
map <- mysterycall_map_leaflet()
}
}
\seealso{
Other mapping: 
\code{\link{mysterycall_clear_isochrone_cache}()},
\code{\link{mysterycall_create_isochrones}()},
\code{\link{mysterycall_isochrones_for_df}()},
\code{\link{mysterycall_map_acceptance_rate}()},
\code{\link{mysterycall_map_acog_districts}()},
\code{\link{mysterycall_map_base}()},
\code{\link{mysterycall_map_block_group}()},
\code{\link{mysterycall_map_physicians}()},
\code{\link{mysterycall_plot_density}()},
\code{\link{mysterycall_plot_isochrones}()},
\code{\link{mysterycall_plot_line}()},
\code{\link{mysterycall_plot_scatter}()}
}
\concept{mapping}
