% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress-bars.R
\name{mysterycall_multi_done}
\alias{mysterycall_multi_done}
\title{Complete multi-step tracker}
\usage{
mysterycall_multi_done(tracker)
}
\arguments{
\item{tracker}{Multi-progress tracker object}
}
\value{
Invisible NULL
}
\description{
Complete multi-step tracker
}
\examples{
\donttest{
tracker <- mysterycall_multi_step(c("Geocode", "Validate"), 10)
mysterycall_multi_done(tracker)
}
}
\seealso{
Other progress-bars: 
\code{\link{mysterycall_multi_complete}()},
\code{\link{mysterycall_multi_progress}()},
\code{\link{mysterycall_multi_step}()},
\code{\link{mysterycall_multi_update}()},
\code{\link{mysterycall_progress_bar}()},
\code{\link{mysterycall_progress_done}()},
\code{\link{mysterycall_progress_fail}()},
\code{\link{mysterycall_progress_map}()},
\code{\link{mysterycall_progress_update}()},
\code{\link{mysterycall_spinner_start}()},
\code{\link{mysterycall_spinner_stop}()}
}
\concept{progress-bars}
