% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysterycall_run_workflow_logged.R
\name{mysterycall_print_dashboard}
\alias{mysterycall_print_dashboard}
\title{Print a formatted summary dashboard}
\usage{
mysterycall_print_dashboard(results)
}
\arguments{
\item{results}{List containing workflow results (as returned by
\code{\link[=mysterycall_run_workflow]{mysterycall_run_workflow()}}).}
}
\description{
Print a formatted summary dashboard
}
\seealso{
Other workflow: 
\code{\link{mysterycall_clean_phase1}()},
\code{\link{mysterycall_clean_phase2}()},
\code{\link{mysterycall_rename_columns}()},
\code{\link{mysterycall_run_workflow}()},
\code{\link{mysterycall_run_workflow_logged}()},
\code{\link{mysterycall_split_and_save}()}
}
\concept{workflow}
