% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress-bars.R
\name{mysterycall_spinner_start}
\alias{mysterycall_spinner_start}
\title{Show a spinner for indeterminate operations}
\usage{
mysterycall_spinner_start(name, msg = NULL)
}
\arguments{
\item{name}{Operation name}

\item{msg}{Optional message to display}
}
\value{
Spinner ID
}
\description{
Shows an animated spinner for operations where total is unknown.
}
\examples{
\donttest{
id <- mysterycall_spinner_start("Loading data")
Sys.sleep(0.1)
mysterycall_spinner_stop(id)
}
}
