% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_generate_overall.R
\name{mysterycall_table_overall}
\alias{mysterycall_table_overall}
\title{Generate an overall summary table}
\usage{
mysterycall_table_overall(
  input_file_path,
  output_directory,
  title = "Overall Table Summary",
  selected_columns = NULL,
  label_translations = NULL
)
}
\arguments{
\item{input_file_path}{The path to an RDS data file.}

\item{output_directory}{The directory where the output table file will be saved.}

\item{title}{The title for the overall table summary (default is "Overall Table Summary").}

\item{selected_columns}{Optional vector of selected columns to include in the table.}

\item{label_translations}{Optional named list for label translations.}
}
\value{
Path to the generated PDF file
}
\description{
Generate an overall table summarizing the demographics of the Table 1.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Generate the overall table
mysterycall_table_overall("data/Table1.rds", "output_tables")
\dontshow{\}) # examplesIf}
}
\seealso{
Other table: 
\code{\link{mysterycall_max_table}()},
\code{\link{mysterycall_min_table}()},
\code{\link{mysterycall_table_percentages}()},
\code{\link{mysterycall_table_proportion}()},
\code{\link{mysterycall_write_arsenal_table}()}
}
\concept{table}
