% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal_tables_write.R
\name{mysterycall_write_arsenal_table}
\alias{mysterycall_write_arsenal_table}
\alias{mysterycall_write_arsenal_table}
\title{Writes an Arsenal table object to a Word document.}
\usage{
mysterycall_write_arsenal_table(object, filename, output_dir = NULL)
}
\arguments{
\item{object}{An object to be written to Word, typically an Arsenal table.}

\item{filename}{The filename (without extension) for the Word document.}

\item{output_dir}{Directory where the Word document should be written.
Defaults to a session-specific folder inside \code{\link[=tempdir]{tempdir()}}.}
}
\value{
Called for its side effect of saving a Word document to \code{word_path}. Returns \code{NULL} invisibly.
}
\description{
Writes an Arsenal table object to a Word document.
}
\examples{
\dontrun{
mysterycall_write_arsenal_table(my_table, "output_table")
}
}
\seealso{
Other table: 
\code{\link{mysterycall_max_table}()},
\code{\link{mysterycall_min_table}()},
\code{\link{mysterycall_table_percentages}()},
\code{\link{mysterycall_table_proportion}()},
\code{\link{mysterycall_table_overall}()}
}
\concept{table}
