% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_indicators.R
\docType{data}
\name{NCI_CANCER_CENTERS}
\alias{NCI_CANCER_CENTERS}
\title{NCI-Designated Cancer Center Patterns}
\format{
Character vector of NCI cancer center text patterns.
}
\usage{
NCI_CANCER_CENTERS
}
\description{
String patterns for National Cancer Institute designated cancer centers.
Confidence: 0.98.
}
\examples{
NCI_CANCER_CENTERS

}
\seealso{
\code{\link{classify_academic_affiliation}}
}
\keyword{datasets}
