% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_normalizer.R
\name{normalize_directionals}
\alias{normalize_directionals}
\title{Normalize Directional Prefixes and Suffixes}
\usage{
normalize_directionals(addr)
}
\arguments{
\item{addr}{Character vector of address strings to normalize.}
}
\value{
Character vector with directionals converted to USPS abbreviations.
}
\description{
Converts full directional words (NORTH, SOUTH, EAST, WEST) and compound
directions (NORTHEAST, etc.) to USPS abbreviations (N, S, E, W, NE, etc.).
}
\examples{
normalize_directionals("123 North Main Street")
normalize_directionals("456 Southeast Oak Avenue")

}
\seealso{
Other address-normalization: 
\code{\link{ascii_norm}()},
\code{\link{caps}()},
\code{\link{has_street_number}()},
\code{\link{is_po_box}()},
\code{\link{normalize_address_df}()},
\code{\link{normalize_state}()},
\code{\link{normalize_suffix}()},
\code{\link{normalize_units}()},
\code{\link{normalize_zip5}()},
\code{\link{strip_suite}()}
}
\concept{address-normalization}
