% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_normalizer.R
\name{normalize_suffix}
\alias{normalize_suffix}
\title{Normalize Street Suffixes}
\usage{
normalize_suffix(addr)
}
\arguments{
\item{addr}{Character vector of address strings to normalize.}
}
\value{
Character vector with street suffixes converted to USPS abbreviations.
}
\description{
Converts full street suffix names (STREET, AVENUE, ROAD, etc.) to their
USPS abbreviations (ST, AVE, RD, etc.).
}
\examples{
normalize_suffix("123 Main Street")
normalize_suffix("456 Oak Boulevard Suite 100")

}
\seealso{
Other address-normalization: 
\code{\link{ascii_norm}()},
\code{\link{caps}()},
\code{\link{has_street_number}()},
\code{\link{is_po_box}()},
\code{\link{normalize_address_df}()},
\code{\link{normalize_directionals}()},
\code{\link{normalize_state}()},
\code{\link{normalize_units}()},
\code{\link{normalize_zip5}()},
\code{\link{strip_suite}()}
}
\concept{address-normalization}
