% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller_reliability.R
\name{print.mysterycall_reliability}
\alias{print.mysterycall_reliability}
\title{Print method for mysterycall_reliability objects}
\usage{
\method{print}{mysterycall_reliability}(x, ...)
}
\arguments{
\item{x}{A \code{mysterycall_reliability} object.}

\item{...}{Ignored.}
}
\value{
\code{invisible(x)}.
}
\description{
Print method for mysterycall_reliability objects
}
\seealso{
Other caller-management: 
\code{\link{mysterycall_caller_reliability}()}
}
\concept{caller-management}
