% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{print.mysterycall_table1}
\alias{print.mysterycall_table1}
\title{Print a mysterycall_table1 object}
\usage{
\method{print}{mysterycall_table1}(x, ...)
}
\arguments{
\item{x}{A \code{mysterycall_table1} object returned by \code{\link[=mysterycall_table1]{mysterycall_table1()}}.}

\item{...}{Additional arguments passed to \code{\link[tibble:formatting]{tibble::print.tbl_df()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a mysterycall_table1 object
}
\seealso{
Other table: 
\code{\link{mysterycall_disparities_table}()},
\code{\link{mysterycall_format_pct}()},
\code{\link{mysterycall_max_table}()},
\code{\link{mysterycall_min_table}()},
\code{\link{mysterycall_model_table}()},
\code{\link{mysterycall_table1}()},
\code{\link{mysterycall_table1_gtsummary}()},
\code{\link{mysterycall_table_overall}()},
\code{\link{mysterycall_table_percentages}()},
\code{\link{mysterycall_table_proportion}()},
\code{\link{mysterycall_write_arsenal_table}()},
\code{\link{mysterycall_write_table_pdf}()},
\code{\link{print.mysterycall_disparities_table}()}
}
\concept{table}
