% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_tracker.R
\name{progress_tracker_fail}
\alias{progress_tracker_fail}
\title{Mark a step as failed}
\usage{
progress_tracker_fail(tracker, step, reason = NULL)
}
\arguments{
\item{tracker}{Object created by \code{\link[=progress_tracker]{progress_tracker()}}.}

\item{step}{Step name.}

\item{reason}{Optional string describing why the step failed.}
}
\description{
Mark a step as failed
}
