% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_tracker.R
\name{progress_tracker_update}
\alias{progress_tracker_update}
\title{Emit a manual progress update}
\usage{
progress_tracker_update(tracker, force = FALSE)
}
\arguments{
\item{tracker}{Object created by \code{\link[=progress_tracker]{progress_tracker()}}.}

\item{force}{Logical flag indicating whether the update should be emitted
even if the configured interval has not elapsed.}
}
\description{
Emit a manual progress update
}
\seealso{
Other logging utilities: 
\code{\link{progress_tracker}()},
\code{\link{progress_tracker_fail}()},
\code{\link{progress_tracker_finish}()},
\code{\link{progress_tracker_start}()},
\code{\link{progress_tracker_summary}()},
\code{\link{tyler_log_cache_hit}()},
\code{\link{tyler_log_error}()},
\code{\link{tyler_log_info}()},
\code{\link{tyler_log_progress}()},
\code{\link{tyler_log_save}()},
\code{\link{tyler_log_step}()},
\code{\link{tyler_log_step_complete}()},
\code{\link{tyler_log_success}()},
\code{\link{tyler_log_warning}()},
\code{\link{tyler_workflow_start}()}
}
\concept{logging utilities}
