% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_constant_vars.R
\name{remove_constant_vars}
\alias{remove_constant_vars}
\title{Remove Constant Variables from a Data Frame}
\usage{
remove_constant_vars(data_frame)
}
\arguments{
\item{data_frame}{A data frame from which constant variables should be removed.}
}
\value{
A data frame with constant variables removed.
}
\description{
This function takes a data frame and returns a new data frame with constant variables removed.
}
\examples{
\dontrun{
new_data <- remove_constant_vars(data_frame)
}

}
\seealso{
Other utilities: 
\code{\link{download_large_file}()},
\code{\link{format_pct}()},
\code{\link{remove_near_zero_var}()},
\code{\link{save_quality_check_table}()},
\code{\link{tyler_assess_data_quality}()},
\code{\link{tyler_check_api_response}()},
\code{\link{tyler_check_data_completeness}()},
\code{\link{tyler_check_dependencies}()},
\code{\link{tyler_check_no_data_loss}()},
\code{\link{tyler_check_no_limits}()},
\code{\link{tyler_estimate_resources}()},
\code{\link{tyler_export_with_backup}()},
\code{\link{tyler_preflight_check}()},
\code{\link{tyler_quality_tier}()},
\code{\link{tyler_resolve_path}()},
\code{\link{tyler_scan_for_limits}()},
\code{\link{tyler_standard_labels}()},
\code{\link{tyler_standard_palette}()},
\code{\link{tyler_use_quiet_logging}()}
}
\concept{utilities}
