% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_near_zero_var.R
\name{remove_near_zero_var}
\alias{remove_near_zero_var}
\title{Remove Near-Zero Variance Variables from a Data Frame}
\usage{
remove_near_zero_var(data_frame, freqCut = 19, uniqueCut = 10)
}
\arguments{
\item{data_frame}{A data frame from which near-zero variance variables should be removed.}

\item{freqCut}{The ratio of the most common value to the second most common value. Defaults to 19.}

\item{uniqueCut}{The percentage of distinct values out of the number of total samples. Defaults to 10.}
}
\value{
A data frame with near-zero variance variables removed.
}
\description{
This function takes a data frame and returns a new data frame with near-zero variance variables removed.
}
\examples{
\dontrun{
new_data <- remove_near_zero_var(data_frame)
}

}
\seealso{
Other utilities: 
\code{\link{download_large_file}()},
\code{\link{format_pct}()},
\code{\link{remove_constant_vars}()},
\code{\link{save_quality_check_table}()},
\code{\link{tyler_assess_data_quality}()},
\code{\link{tyler_check_api_response}()},
\code{\link{tyler_check_data_completeness}()},
\code{\link{tyler_check_dependencies}()},
\code{\link{tyler_check_no_data_loss}()},
\code{\link{tyler_check_no_limits}()},
\code{\link{tyler_estimate_resources}()},
\code{\link{tyler_export_with_backup}()},
\code{\link{tyler_preflight_check}()},
\code{\link{tyler_quality_tier}()},
\code{\link{tyler_resolve_path}()},
\code{\link{tyler_scan_for_limits}()},
\code{\link{tyler_standard_labels}()},
\code{\link{tyler_standard_palette}()},
\code{\link{tyler_use_quiet_logging}()}
}
\concept{utilities}
