% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phase_2_results.R
\name{rename_columns_by_substring}
\alias{rename_columns_by_substring}
\title{Rename columns based on substring matches}
\usage{
rename_columns_by_substring(data, target_strings, new_names)
}
\arguments{
\item{data}{A data frame whose columns need renaming.}

\item{target_strings}{A vector of substrings to search for within column names.}

\item{new_names}{A vector of new names corresponding to the target strings.}
}
\value{
A data frame with renamed columns.
}
\description{
This function searches through column names in a data frame for specified substrings and renames the first matching column to a new name provided by the user. It provides detailed logs for each operation, including the columns found and any renaming actions taken. If multiple columns match a substring, only the first is renamed, and a warning is issued.
}
\examples{
df <- data.frame(
  doctor_info = 1:5,
  patient_contact_data = 6:10
)
# Renaming 'doctor_info' to 'physician_info'
df <- rename_columns_by_substring(df,
                                  target_strings = c("doctor"),
                                  new_names = c("physician_info"))
print(df)

# More complex example with multiple renamings
df <- data.frame(
  doc_information = 1:5,
  patient_contact = 6:10,
  doctor_notes = 11:15
)
# Renaming 'doc_information' to 'doctor_info' and 'doctor_notes' to 'notes'
df <- rename_columns_by_substring(df,
                                  target_strings = c("doc_information", "doctor_notes"),
                                  new_names = c("doctor_info", "notes"))
print(df)
}
