% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinician_data.R
\name{retrieve_clinician_data}
\alias{retrieve_clinician_data}
\title{Retrieve Clinician Data}
\usage{
retrieve_clinician_data(input_data)
}
\arguments{
\item{input_data}{Either a dataframe containing NPI numbers or a path to a CSV file.}
}
\value{
A tibble with clinician data for the provided NPIs.
}
\description{
This function retrieves clinician data for each valid NPI in the input dataframe.
}
