% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_journal.R
\name{scale_fill_green_journal}
\alias{scale_fill_green_journal}
\title{Green Journal discrete fill scale}
\usage{
scale_fill_green_journal(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 \link[ggplot2:Scale]{ggplot2::Scale} object.
}
\description{
Convenience wrapper for \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}} using the
Okabe-Ito palette from \code{\link[=palette_green_journal]{palette_green_journal()}}.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, fill = drv)) +
  geom_bar() +
  scale_fill_green_journal()
}
\seealso{
Other green-journal-colors: 
\code{\link{palette_green_journal}()},
\code{\link{scale_color_green_journal}()}
}
\concept{green-journal-colors}
