% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this_one_works.R
\name{scrape_physicians_data}
\alias{scrape_physicians_data}
\title{Scrape Physicians' Data}
\usage{
scrape_physicians_data(startID, endID)
}
\arguments{
\item{startID}{The starting ID for scraping.}

\item{endID}{The ending ID for scraping.}
}
\value{
A dataframe containing scraped physicians' data.
}
\description{
This function scrapes data for physicians within a specified ID range, excluding wrong IDs.
}
\examples{
# Call the function
scrape_result <- scrape_physicians_data(startID = 9045999, endID = 9046000)

}
