% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_by_taxonomy.R
\name{search_by_taxonomy}
\alias{search_by_taxonomy}
\title{Search NPI Database by Taxonomy}
\usage{
search_by_taxonomy(taxonomy_to_search)
}
\arguments{
\item{taxonomy_to_search}{A character vector containing the taxonomy description(s) to search for.}
}
\value{
A data frame with filtered NPI data based on the specified taxonomy description.
}
\description{
This function searches the NPI Database for healthcare providers based on a taxonomy description.
}
\examples{
# Example usage with multiple taxonomy descriptions:
data <- search_by_taxonomy(c("Gynecologic Oncology", "Female Pelvic Medicine and Reconstructive Surgery", "Reproductive Endocrinology", "Maternal & Fetal Medicine"))

}
