% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitting_dataframe_to_send_to_callers.R
\name{split_and_save}
\alias{split_and_save}
\title{This function takes an input data file, splits it into multiple parts, and saves each part as a separate Excel file.}
\usage{
split_and_save(input_path, output_directory, lab_assistant_names)
}
\arguments{
\item{input_path}{The path to the input data file (RDS, CSV, or XLS/XLSX).}

\item{output_directory}{The directory where the output Excel files will be saved.}

\item{lab_assistant_names}{A vector of lab_assistant_names to name the output files.}
}
\value{
None
}
\description{
This function takes an input data file, splits it into multiple parts, and saves each part as a separate Excel file.
}
\examples{
\dontrun{
library(tyler)
input_path <- "/path/to/your/input/file.rds"
output_directory <- "/path/to/your/output/directory"
lab_assistant_names <- c("Label1", "Label2", "Label3")
split_and_save(input_path, output_directory, lab_assistant_names)
}
}
