% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{ACGME OBGYN Residency Data}
\format{
Data frame with columns
\describe{
\item{NUCC data}{taxonomy.}{taxonomy codes.}
\item{name}
}
}
\source{
\url{https://www.nucc.org/images/stories/PDF/taxonomy_23_0.pdf}
}
\usage{
taxonomy
}
\description{
This data set contains taxonomy codes for Obstetricians and Gynecologists among others.
}
\examples{
  taxonomy
}
\keyword{datasets}
