% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tyler-deprecated}
\alias{tyler-deprecated}
\alias{tyler_assess_data_quality}
\alias{tyler_calculate_overlap}
\alias{tyler_check_api_response}
\alias{tyler_check_data_completeness}
\alias{tyler_check_dependencies}
\alias{tyler_check_no_data_loss}
\alias{tyler_check_no_limits}
\alias{tyler_check_normality}
\alias{tyler_clean_phase1}
\alias{tyler_clean_phase2}
\alias{tyler_clear_isochrone_cache}
\alias{tyler_create_formula}
\alias{tyler_create_isochrones}
\alias{tyler_download_file}
\alias{tyler_estimate_resources}
\alias{tyler_export_with_backup}
\alias{tyler_format_duration}
\alias{tyler_format_pct}
\alias{tyler_genderize}
\alias{tyler_geocode}
\alias{tyler_get_census_data}
\alias{tyler_get_clinician_data}
\alias{tyler_hrr}
\alias{tyler_hrr_maps}
\alias{tyler_isochrones_for_df}
\alias{tyler_log_cache_hit}
\alias{tyler_log_error}
\alias{tyler_log_info}
\alias{tyler_log_progress}
\alias{tyler_log_save}
\alias{tyler_log_step}
\alias{tyler_log_step_complete}
\alias{tyler_log_success}
\alias{tyler_log_warning}
\alias{tyler_map_acog_districts}
\alias{tyler_map_base}
\alias{tyler_map_block_group}
\alias{tyler_map_leaflet}
\alias{tyler_map_physicians}
\alias{tyler_max_table}
\alias{tyler_min_table}
\alias{tyler_most_common_gender}
\alias{tyler_multi_complete}
\alias{tyler_multi_done}
\alias{tyler_multi_progress}
\alias{tyler_multi_step}
\alias{tyler_multi_update}
\alias{tyler_not_contacted_states}
\alias{tyler_physician_age}
\alias{tyler_plot_census_age}
\alias{tyler_plot_density}
\alias{tyler_plot_emmeans}
\alias{tyler_plot_interaction}
\alias{tyler_plot_isochrones}
\alias{tyler_plot_line}
\alias{tyler_plot_scatter}
\alias{tyler_preflight_check}
\alias{tyler_print_dashboard}
\alias{tyler_progress_bar}
\alias{tyler_progress_callback}
\alias{tyler_progress_done}
\alias{tyler_progress_fail}
\alias{tyler_progress_finish}
\alias{tyler_progress_map}
\alias{tyler_progress_start}
\alias{tyler_progress_summary}
\alias{tyler_progress_tracker}
\alias{tyler_progress_update}
\alias{tyler_quality_tier}
\alias{tyler_remove_constants}
\alias{tyler_remove_near_zero}
\alias{tyler_rename_columns}
\alias{tyler_resolve_path}
\alias{tyler_run_workflow}
\alias{tyler_run_workflow_logged}
\alias{tyler_save_quality_table}
\alias{tyler_scan_for_limits}
\alias{tyler_search_and_process_npi}
\alias{tyler_search_taxonomy}
\alias{tyler_spinner_start}
\alias{tyler_spinner_stop}
\alias{tyler_split_and_save}
\alias{tyler_standard_labels}
\alias{tyler_standard_palette}
\alias{tyler_summarize_census}
\alias{tyler_table_overall}
\alias{tyler_table_percentages}
\alias{tyler_table_proportion}
\alias{tyler_use_quiet_logging}
\alias{tyler_validate_npi}
\alias{tyler_workflow_end}
\alias{tyler_workflow_start}
\alias{tyler_write_arsenal_table}
\alias{tyler_write_table_pdf}
\title{Deprecated tyler_ prefix functions}
\usage{
tyler_assess_data_quality(...)

tyler_calculate_overlap(...)

tyler_check_api_response(...)

tyler_check_data_completeness(...)

tyler_check_dependencies(...)

tyler_check_no_data_loss(...)

tyler_check_no_limits(...)

tyler_check_normality(...)

tyler_clean_phase1(...)

tyler_clean_phase2(...)

tyler_clear_isochrone_cache(...)

tyler_create_formula(...)

tyler_create_isochrones(...)

tyler_download_file(...)

tyler_estimate_resources(...)

tyler_export_with_backup(...)

tyler_format_duration(...)

tyler_format_pct(...)

tyler_genderize(...)

tyler_geocode(...)

tyler_get_census_data(...)

tyler_get_clinician_data(...)

tyler_hrr(...)

tyler_hrr_maps(...)

tyler_isochrones_for_df(...)

tyler_log_cache_hit(...)

tyler_log_error(...)

tyler_log_info(...)

tyler_log_progress(...)

tyler_log_save(...)

tyler_log_step(...)

tyler_log_step_complete(...)

tyler_log_success(...)

tyler_log_warning(...)

tyler_map_acog_districts(...)

tyler_map_base(...)

tyler_map_block_group(...)

tyler_map_leaflet(...)

tyler_map_physicians(...)

tyler_max_table(...)

tyler_min_table(...)

tyler_most_common_gender(...)

tyler_multi_complete(...)

tyler_multi_done(...)

tyler_multi_progress(...)

tyler_multi_step(...)

tyler_multi_update(...)

tyler_not_contacted_states(...)

tyler_physician_age(...)

tyler_plot_census_age(...)

tyler_plot_density(...)

tyler_plot_emmeans(...)

tyler_plot_interaction(...)

tyler_plot_isochrones(...)

tyler_plot_line(...)

tyler_plot_scatter(...)

tyler_preflight_check(...)

tyler_print_dashboard(...)

tyler_progress_bar(...)

tyler_progress_callback(...)

tyler_progress_done(...)

tyler_progress_fail(...)

tyler_progress_finish(...)

tyler_progress_map(...)

tyler_progress_start(...)

tyler_progress_summary(...)

tyler_progress_tracker(...)

tyler_progress_update(...)

tyler_quality_tier(...)

tyler_remove_constants(...)

tyler_remove_near_zero(...)

tyler_rename_columns(...)

tyler_resolve_path(...)

tyler_run_workflow(...)

tyler_run_workflow_logged(...)

tyler_save_quality_table(...)

tyler_scan_for_limits(...)

tyler_search_and_process_npi(...)

tyler_search_taxonomy(...)

tyler_spinner_start(...)

tyler_spinner_stop(...)

tyler_split_and_save(...)

tyler_standard_labels(...)

tyler_standard_palette(...)

tyler_summarize_census(...)

tyler_table_overall(...)

tyler_table_percentages(...)

tyler_table_proportion(...)

tyler_use_quiet_logging(...)

tyler_validate_npi(...)

tyler_workflow_end(...)

tyler_workflow_start(...)

tyler_write_arsenal_table(...)

tyler_write_table_pdf(...)
}
\description{
These functions existed when the package was named \code{tyler}. They have been
renamed with the \code{mysterycall_} prefix. All \verb{tyler_*} wrappers call their
\verb{mysterycall_*} counterpart and emit a deprecation warning.
}
\keyword{internal}
