% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler-package.R
\docType{package}
\name{tyler-package}
\alias{tyler}
\alias{tyler-package}
\title{tyler: Tools for Mystery Caller and Audit Healthcare Studies}
\description{
Provides utilities for mystery caller and audit studies evaluating patient access to healthcare. Functions help search and process National Provider Identifier (NPI) records, summarize provider demographics, and generate tables and maps for analysis and reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mufflyt/tyler}
  \item \url{https://mufflyt.github.io/tyler/}
  \item Report bugs at \url{https://github.com/mufflyt/tyler/issues}
}

}
\author{
\strong{Maintainer}: Tyler Muffly \email{tyler.muffly@dhha.org} (ORCID: 0000-0002-2044-1693)

}
\keyword{NPI}
\keyword{NPPES}
\keyword{healthcare}
\keyword{internal}
