% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cache.R
\name{tyler_cache_dir}
\alias{tyler_cache_dir}
\title{Determine the cache directory used for downloaded resources}
\usage{
tyler_cache_dir(...)
}
\arguments{
\item{...}{Optional path components appended to the cache directory.}
}
\value{
A character string with the cache directory path.
}
\description{
Determine the cache directory used for downloaded resources
}
