% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_checks.R
\name{tyler_estimate_resources}
\alias{tyler_estimate_resources}
\title{Estimate workflow resources}
\usage{
tyler_estimate_resources(n_rows)
}
\arguments{
\item{n_rows}{Integer number of input rows used to project runtime and
memory requirements.}
}
\value{
List with runtime and memory estimates
}
\description{
Estimate workflow resources
}
