% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_log_cache_hit}
\alias{tyler_log_cache_hit}
\title{Log cache hit}
\usage{
tyler_log_cache_hit(what, n_items)
}
\arguments{
\item{what}{What was loaded from cache}

\item{n_items}{Number of cached items}
}
\value{
\code{invisible(NULL)}.
}
\description{
Log cache hit
}
\seealso{
Other logging utilities: 
\code{\link{progress_tracker}()},
\code{\link{progress_tracker_fail}()},
\code{\link{progress_tracker_finish}()},
\code{\link{progress_tracker_start}()},
\code{\link{progress_tracker_summary}()},
\code{\link{progress_tracker_update}()},
\code{\link{tyler_log_error}()},
\code{\link{tyler_log_info}()},
\code{\link{tyler_log_progress}()},
\code{\link{tyler_log_save}()},
\code{\link{tyler_log_step}()},
\code{\link{tyler_log_step_complete}()},
\code{\link{tyler_log_success}()},
\code{\link{tyler_log_warning}()},
\code{\link{tyler_workflow_start}()}
}
\concept{logging utilities}
