% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_log_save}
\alias{tyler_log_save}
\title{Log file save}
\usage{
tyler_log_save(path, n_rows = NULL)
}
\arguments{
\item{path}{File path}

\item{n_rows}{Number of rows (optional)}
}
\value{
Invisible NULL
}
\description{
Log file save
}
\seealso{
Other logging utilities: 
\code{\link{tyler_log_cache_hit}()},
\code{\link{tyler_log_error}()},
\code{\link{tyler_log_info}()},
\code{\link{tyler_log_progress}()},
\code{\link{tyler_log_step}()},
\code{\link{tyler_log_step_complete}()},
\code{\link{tyler_log_success}()},
\code{\link{tyler_log_warning}()},
\code{\link{tyler_workflow_start}()}
}
\concept{logging utilities}
