% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler_map_block_group.R
\name{tyler_map_block_group}
\alias{tyler_map_block_group}
\alias{tyler_map_block_group}
\title{Function to create and export a map showing block group overlap with isochrones}
\usage{
tyler_map_block_group(
  bg_data,
  isochrones_data,
  output_dir = "figures/"
)
}
\arguments{
\item{bg_data}{A SpatialPolygonsDataFrame representing block group data.}

\item{isochrones_data}{A SpatialPolygonsDataFrame representing isochrone data.}

\item{output_dir}{Directory path for exporting the map files. Default is "figures/".}
}
\value{
Called for its side effects of saving the block group overlap map as HTML and PNG inside \code{output_dir}. Returns \code{NULL} invisibly.
}
\description{
This function creates a map that displays block groups and their overlap with isochrones.
The map is exported as an HTML file and a PNG image.
}
\examples{
\dontrun{
# Create and export the map with the default output directory
tyler_map_block_group(block_groups, isochrones_joined_map)

# Create and export the map with a custom output directory
tyler_map_block_group(block_groups, isochrones_joined_map, "custom_output/")
}

}
\seealso{
Other mapping: 
\code{\link{tyler_plot_density}()},
\code{\link{tyler_plot_isochrones}()},
\code{\link{tyler_create_isochrones}()},
\code{\link{tyler_isochrones_for_df}()},
\code{\link{tyler_plot_line}()},
\code{\link{tyler_plot_scatter}()},
\code{\link{tyler_map_acog_districts}()},
\code{\link{tyler_map_base}()},
\code{\link{tyler_map_leaflet}()},
\code{\link{tyler_map_physicians}()}
}
\concept{mapping}
