% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler_map_leaflet.R
\name{tyler_map_leaflet}
\alias{tyler_map_leaflet}
\alias{tyler_map_leaflet}
\title{Create a Leaflet Base Map}
\usage{
tyler_map_leaflet()
}
\value{
Invisibly returns the Leaflet map object.
}
\description{
This function creates a Leaflet BASE map with specific configurations, including the base tile layer, scale bar,
default view settings, and layers control.
}
\examples{
\donttest{
map <- tyler_map_leaflet()
}
}
\seealso{
Other mapping: 
\code{\link{tyler_plot_density}()},
\code{\link{tyler_plot_isochrones}()},
\code{\link{tyler_create_isochrones}()},
\code{\link{tyler_isochrones_for_df}()},
\code{\link{tyler_plot_line}()},
\code{\link{tyler_plot_scatter}()},
\code{\link{tyler_map_acog_districts}()},
\code{\link{tyler_map_base}()},
\code{\link{tyler_map_block_group}()},
\code{\link{tyler_map_physicians}()}
}
\concept{mapping}
