% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress-bars.R
\name{tyler_multi_complete}
\alias{tyler_multi_complete}
\title{Complete current step in multi-progress tracker}
\usage{
tyler_multi_complete(tracker, result = NULL)
}
\arguments{
\item{tracker}{Multi-progress tracker object}

\item{result}{Optional result message}
}
\value{
Invisible NULL
}
\description{
Complete current step in multi-progress tracker
}
