% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress-bars.R
\name{tyler_multi_done}
\alias{tyler_multi_done}
\title{Complete multi-step tracker}
\usage{
tyler_multi_done(tracker)
}
\arguments{
\item{tracker}{Multi-progress tracker object}
}
\value{
Invisible NULL
}
\description{
Complete multi-step tracker
}
