% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler_progress_tracker.R
\name{tyler_progress_update}
\alias{tyler_progress_update}
\alias{tyler_progress_update}
\title{Emit a manual progress update}
\usage{
tyler_progress_update(tracker, force = FALSE)
}
\arguments{
\item{tracker}{Object created by \code{\link[=tyler_progress_tracker]{tyler_progress_tracker()}}.}

\item{force}{Logical flag indicating whether the update should be emitted
even if the configured interval has not elapsed.}
}
\description{
Emit a manual progress update
}
\seealso{
Other logging utilities: 
\code{\link{tyler_progress_tracker}()},
\code{\link{tyler_progress_fail}()},
\code{\link{tyler_progress_finish}()},
\code{\link{tyler_progress_start}()},
\code{\link{tyler_progress_summary}()},
\code{\link{tyler_log_cache_hit}()},
\code{\link{tyler_log_error}()},
\code{\link{tyler_log_info}()},
\code{\link{tyler_log_progress}()},
\code{\link{tyler_log_save}()},
\code{\link{tyler_log_step}()},
\code{\link{tyler_log_step_complete}()},
\code{\link{tyler_log_success}()},
\code{\link{tyler_log_warning}()},
\code{\link{tyler_workflow_start}()}
}
\concept{logging utilities}
