% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler_remove_constants.R
\name{tyler_remove_constants}
\alias{tyler_remove_constants}
\alias{tyler_remove_constants}
\title{Remove Constant Variables from a Data Frame}
\usage{
tyler_remove_constants(data_frame)
}
\arguments{
\item{data_frame}{A data frame from which constant variables should be removed.}
}
\value{
A data frame with constant variables removed.
}
\description{
This function takes a data frame and returns a new data frame with constant variables removed.
}
\examples{
\dontrun{
new_data <- tyler_remove_constants(data_frame)
}

}
\seealso{
Other utilities: 
\code{\link{tyler_download_file}()},
\code{\link{tyler_format_pct}()},
\code{\link{tyler_remove_near_zero}()},
\code{\link{tyler_save_quality_table}()},
\code{\link{tyler_assess_data_quality}()},
\code{\link{tyler_check_api_response}()},
\code{\link{tyler_check_data_completeness}()},
\code{\link{tyler_check_dependencies}()},
\code{\link{tyler_check_no_data_loss}()},
\code{\link{tyler_check_no_limits}()},
\code{\link{tyler_estimate_resources}()},
\code{\link{tyler_export_with_backup}()},
\code{\link{tyler_preflight_check}()},
\code{\link{tyler_quality_tier}()},
\code{\link{tyler_resolve_path}()},
\code{\link{tyler_scan_for_limits}()},
\code{\link{tyler_standard_labels}()},
\code{\link{tyler_standard_palette}()},
\code{\link{tyler_use_quiet_logging}()}
}
\concept{utilities}
