% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress-bars.R
\name{tyler_spinner_stop}
\alias{tyler_spinner_stop}
\title{Stop a spinner}
\usage{
tyler_spinner_stop(id, result = "done")
}
\arguments{
\item{id}{Spinner ID from tyler_spinner_start()}

\item{result}{Result message}
}
\value{
Invisible NULL
}
\description{
Stop a spinner
}
