% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler_table_overall.R
\name{tyler_table_overall}
\alias{tyler_table_overall}
\alias{tyler_table_overall}
\title{Generate an overall summary table}
\usage{
tyler_table_overall(
  input_file_path,
  output_directory,
  title = "Overall Table Summary",
  selected_columns = NULL,
  label_translations = NULL
)
}
\arguments{
\item{input_file_path}{The path to an RDS data file.}

\item{output_directory}{The directory where the output table file will be saved.}

\item{title}{The title for the overall table summary (default is "Overall Table Summary").}

\item{selected_columns}{Optional vector of selected columns to include in the table.}

\item{label_translations}{Optional named list for label translations.}
}
\value{
Path to the generated PDF file
}
\description{
Generate an overall table summarizing the demographics of the Table 1.
}
\examples{
\dontrun{
# Generate the overall table
tyler_table_overall("data/Table1.rds", "output_tables")
}
}
\seealso{
Other table: 
\code{\link{tyler_write_arsenal_table}()},
\code{\link{tyler_max_table}()},
\code{\link{tyler_min_table}()},
\code{\link{tyler_table_percentages}()},
\code{\link{tyler_table_proportion}()}
}
\concept{table}
