% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_checks.R
\name{tyler_validate_google_api}
\alias{tyler_validate_google_api}
\title{Validate Google Maps API key}
\usage{
tyler_validate_google_api(api_key)
}
\arguments{
\item{api_key}{Google Maps API key}
}
\value{
List with valid (logical) and error (character) fields
}
\description{
Validate Google Maps API key
}
\keyword{internal}
