% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_checks.R
\name{tyler_validate_here_api}
\alias{tyler_validate_here_api}
\title{Validate HERE API key}
\usage{
tyler_validate_here_api(api_key)
}
\arguments{
\item{api_key}{HERE API key}
}
\value{
List with valid (logical) and error (character) fields
}
\description{
Validate HERE API key
}
\keyword{internal}
