% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler_validate_npi.R
\name{tyler_validate_npi}
\alias{tyler_validate_npi}
\alias{tyler_validate_npi}
\title{Validate and Remove Invalid NPI Numbers}
\usage{
tyler_validate_npi(input_data)
}
\arguments{
\item{input_data}{Either a data frame containing NPI numbers or a path to a CSV file.}
}
\value{
A data frame containing valid NPI numbers.
}
\description{
This function reads a CSV file containing NPI numbers, validates their
format using the npi package, and removes rows with missing or invalid NPIs.
}
