% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_workflow_end}
\alias{tyler_workflow_end}
\title{End workflow and print summary}
\usage{
tyler_workflow_end(final_n = NULL, input_n = NULL)
}
\arguments{
\item{final_n}{Number of final output rows}

\item{input_n}{Number of input rows}
}
\value{
Invisible NULL
}
\description{
End workflow and print summary
}
