% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\docType{data}
\name{us_fips_list}
\alias{us_fips_list}
\title{Get Census data for all states' block groups}
\format{
An object of class \code{character} of length 50.
}
\usage{
us_fips_list
}
\arguments{
\item{us_fips}{A vector of state FIPS codes.}

\item{census_key}{The API key for the getCensus function.}

\item{vintage}{The vintage year for Census data (default is 2019).}
}
\value{
A dataframe containing Census data for all states' block groups.
}
\description{
This function retrieves Census data for all states' block groups by looping
over the specified list of state FIPS codes.
}
\keyword{datasets}
