% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acgme.R
\docType{data}
\name{acgme}
\alias{acgme}
\title{ACGME OBGYN Residency Data}
\format{
Data frame with columns
\describe{
\item{ACGME data}{OBGYN Residency.}{Program directors.}
\item{name}
}
}
\source{
\url{https://apps.acgme.org/ads/Public/Programs/Search}
}
\usage{
acgme
}
\description{
This data set contains information about Obstetricians and Gynecologists residencies.
}
\examples{
  acgme
}
\keyword{datasets}
