% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACOG_Districts.R
\docType{data}
\name{ACOG_Districts}
\alias{ACOG_Districts}
\title{ACOG Districts Data}
\format{
Data frame with columns
\describe{
\item{ACOG_District}{Two letter state abbreviation.}{Full name state.}
\item{name}
}
}
\source{
\url{https://www.acog.org/community/districts-and-sections}
}
\usage{
ACOG_Districts
}
\description{
This data set contains information about American College of Obstetricians and Gynecologists districts.
}
\examples{
  ACOG_Districts
}
\keyword{datasets}
