% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACOG_Districts.R
\docType{data}
\name{ACOG_Districts}
\alias{ACOG_Districts}
\title{ACOG Districts Data}
\format{
A data frame with the following columns:
\describe{
\item{ACOG_District}{Two-letter state abbreviations representing ACOG districts.}
\item{name}{Full names of the states corresponding to the ACOG districts.}
}
}
\source{
Data was obtained from the official ACOG website: \url{https://www.acog.org/community/districts-and-sections}
}
\usage{
ACOG_Districts
}
\description{
This dataset contains information about American College of Obstetricians and Gynecologists (ACOG) districts, including their two-letter state abbreviations and full state names.
}
\examples{
# Load the ACOG Districts Data
data(ACOG_Districts)

# View the first few rows of the dataset
head(ACOG_Districts)

# Get a summary of the dataset
summary(ACOG_Districts)

# Perform data analysis and exploration

}
\keyword{dataset}
