% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal_tables_write.R
\name{arsenal_tables_write2word}
\alias{arsenal_tables_write2word}
\title{Save Arsenal Table as Word Document}
\usage{
arsenal_tables_write2word(object, filename)
}
\arguments{
\item{object}{An Arsenal Table object.}

\item{filename}{A string representing the desired filename without extension.}
}
\description{
This function takes an Arsenal Table object and saves it as a Word document in the "tables" directory.
}
\examples{
\dontrun{
overall <- summary(
  overall_arsenal_table,
  text = T,
  title = 'Table: Characteristics of Obstetrics and Gynecology Subspecialists Practicing at Obstetrics and Gynecology Residency Programs',
  pfootnote = FALSE
)
arsenal_tables_write2word(overall, "arsenal_overall_table_one")
}
}
