% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal_tables_write.R
\name{arsenal_tables_write2word}
\alias{arsenal_tables_write2word}
\title{Writes an Arsenal table object to a Word document.}
\usage{
arsenal_tables_write2word(object, filename)
}
\arguments{
\item{object}{An object to be written to Word, typically an Arsenal table.}

\item{filename}{The filename (without extension) for the Word document.}
}
\value{
None
}
\description{
Writes an Arsenal table object to a Word document.
}
\examples{
\dontrun{
arsenal_tables_write2word(my_table, "output_table")
}
}
