% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate Proportion and Generate Tabulation}
\usage{
calculate_proportion(df, variable_name)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{variable_name}{The name of the categorical variable for which to calculate the proportion.}
}
\value{
A list containing the calculated most common value, proportion variable, and the tabulation result.
}
\description{
This function calculates the proportion of each category in a specified variable
and generates a tabulation of the counts and proportions.
}
\examples{
\dontrun{
library(tyler)
df <- data.frame(gender = c("Male", "Female", "Male", "Male", "Female"))
result <- calculate_proportion(df, "gender")
cat("Most Common Value: ", result$most_common, "\n")
cat("Proportion Variable: ", result$proportion_variable, "\%\n")
print(result$tabyl_result)
}

}
