% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_block_group_overlap_map.R
\name{create_block_group_overlap_map}
\alias{create_block_group_overlap_map}
\title{Function to create and export a map showing block group overlap with isochrones}
\usage{
create_block_group_overlap_map(
  bg_data,
  isochrones_data,
  output_html,
  output_png
)
}
\arguments{
\item{bg_data}{A SpatialPolygonsDataFrame representing block group data.}

\item{isochrones_data}{A SpatialPolygonsDataFrame representing isochrone data.}

\item{output_html}{File path for exporting the map as an HTML file.}

\item{output_png}{File path for exporting the map as a PNG image.}
}
\value{
None
}
\description{
This function creates a map that displays block groups and their overlap with isochrones.
}
\examples{
\dontrun{
# Define output file paths
output_html <- "figures/overlap_bg_map.html"
output_png <- "figures/overlap_bg_map.png"

# Create and export the map
create_block_group_overlap_map(block_groups, isochrones_joined_map, output_html, output_png)

# Call the create_block_group_overlap_map function with your data
create_block_group_overlap_map(
  gyn_onc_physicians_df = centers_ms,
  isochrones_sf = isochrones_combined,
  file_name = html_file_name)
}

}
