% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_individual_isochrone_plots.R
\name{create_individual_isochrone_plots}
\alias{create_individual_isochrone_plots}
\title{Create Individual Isochrone Maps and Shapefiles}
\usage{
create_individual_isochrone_plots(isochrones, drive_times)
}
\arguments{
\item{isochrones}{An sf object containing isochrone data.}

\item{drive_times}{A vector of unique drive times (in minutes) for which maps and shapefiles will be created.}
}
\value{
None. The function creates and saves individual maps and shapefiles.
}
\description{
This function creates individual Leaflet maps and shapefiles for specified drive times
based on isochrone data.
}
\examples{
\dontrun{
# Load required libraries
library(sf)
library(leaflet)
library(tyler)

# Load isochrone data
isochrones <- readRDS("path_to_isochrones.rds")

# List of unique drive times for which you want to create plots and shapefiles
drive_times <- unique(isochrones$drive_time)

# Create individual isochrone maps and shapefiles
create_individual_isochrone_plots(isochrones, drive_times)
}

}
