% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_isochrones.R
\name{create_isochrones}
\alias{create_isochrones}
\title{Memoized function to try a location with isoline calculations}
\usage{
create_isochrones(
  location,
  range,
  posix_time = as.POSIXct("2023-10-20 08:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")
)
}
\arguments{
\item{location}{An sf object representing the location for which isolines will be calculated.}

\item{range}{A numeric vector of time ranges in seconds.}

\item{posix_time}{A POSIXct object representing the date and time of calculation. Default is "2023-10-20 08:00:00".}
}
\value{
A list of isolines for different time ranges, or an error message if the calculation fails.
}
\description{
This function calculates isolines for a given location using the hereR package.
}
\examples{
\dontrun{

# Set your HERE API key in your Renviron file using the following steps:
# 1. Add key to .Renviron
Sys.setenv(HERE_API_KEY = "your_api_key_here")
# 2. Reload .Renviron
readRenviron("~/.Renviron")

# Define a sf object for the location
location <- sf::st_point(c(-73.987, 40.757))

# Calculate isolines for the location with a 30-minute, 60-minute, 120-minute, and 180-minute range
isolines <- create_isochrones(location = location, range = c(1800, 3600, 7200, 10800))

# Print the isolines
print(isolines)

}
}
