% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_isochrones_for_dataframe.R
\name{create_isochrones_for_dataframe}
\alias{create_isochrones_for_dataframe}
\title{Get isochrones for each point in a dataframe}
\usage{
create_isochrones_for_dataframe(dataframe, breaks = c(0, 30, 60, 120, 180))
}
\arguments{
\item{dataframe}{A dataframe containing the points for which isochrones are to be retrieved.}

\item{breaks}{A vector of breaks for labeling the isochrones.}
}
\value{
A dataframe containing the isochrones data with added 'name' column.
}
\description{
This function retrieves isochrones for each point in a given dataframe by looping
over the rows and calling the create_isochrones function for each point.
}
