% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_isochrones_for_dataframe.R
\name{create_isochrones_for_dataframe}
\alias{create_isochrones_for_dataframe}
\title{Get isochrones for each point in a dataframe}
\usage{
create_isochrones_for_dataframe(
  input_file,
  breaks = c(1800, 3600, 7200, 10800)
)
}
\arguments{
\item{input_file}{A path to the input file containing points for which isochrones are to be retrieved.}

\item{breaks}{A numeric vector specifying the breaks for categorizing drive times (default is c(1800, 3600, 7200, 10800)).}
}
\value{
A dataframe containing the isochrones data with added 'name' column.
}
\description{
This function retrieves isochrones for each point in a given dataframe by looping
over the rows and calling the create_isochrones function for each point.
}
