% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\docType{data}
\name{fips}
\alias{fips}
\title{Data of FIPS codes}
\format{
A tibble with 243 rows and 10 variables:
\describe{
}
}
\source{
\url{https://github.com/kjhealy/fips-codes/blob/master/state_and_county_fips_master.csv}
}
\usage{
fips
}
\description{
3 columns with 51 observations.
}
\keyword{datasets}
