% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_by_acog_districts.R
\name{generate_acog_districts_sf}
\alias{generate_acog_districts_sf}
\title{Generate ACOG Districts sf Object}
\usage{
generate_acog_districts_sf()
}
\arguments{
\item{filepath}{The path to the CSV file containing ACOG Districts data.}
}
\value{
An sf object containing grouped ACOG districts.
}
\description{
This function generates an sf object representing the ACOG districts by grouping all the states in a particular district together.
}
