% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overall_table.R
\name{generate_overall_table}
\alias{generate_overall_table}
\title{Generate overall table}
\usage{
generate_overall_table(input_file, output_dir)
}
\arguments{
\item{input_file}{The path to the data file (in RDS, CSV, or XLS format).}

\item{output_dir}{The directory where the output table file will be saved.}
}
\description{
Generate an overall table summarizing the demographics of the Table 1.
}
\examples{
\dontrun{
# Generate the overall table
generate_overall_table("data/Table1.rds", "output_tables")
}
}
