% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_unique_addresses}
\alias{geocode_unique_addresses}
\title{Geocode Unique Addresses from a CSV File}
\usage{
geocode_unique_addresses(file_path, google_maps_api_key, output_file_path)
}
\arguments{
\item{file_path}{The path to the CSV file containing the addresses. The file should have a column named "address".}

\item{google_maps_api_key}{Your Google Maps API key for geocoding.}

\item{output_file_path}{The path where the output CSV file with geocoded data will be saved.}
}
\value{
A data frame containing the original data along with latitude and longitude for each unique address.
}
\description{
This function reads a CSV file containing addresses, geocodes unique addresses using Google Maps API,
and writes the geocoded data back to a CSV file.
}
\examples{
\dontrun{
google_maps_api_key <- "your_google_maps_api_key"
output_data <- geocode_unique_addresses("path/to/your/input/file.csv", google_maps_api_key, "path/to/your/output/file.csv")
}

}
