% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\name{get_census_data}
\alias{get_census_data}
\title{Get Census data for all states' block groups}
\usage{
get_census_data(us_fips, census_key, vintage = 2019)
}
\arguments{
\item{us_fips}{A vector of state FIPS codes.}

\item{census_key}{The API key for the Census API.}

\item{vintage}{The vintage year for Census data (default is 2019).}
}
\value{
A dataframe containing Census data for all states' block groups.

A dataframe containing Census data for all states' block groups.
}
\description{
This function retrieves Census data for all states' block groups by looping
over the specified list of state FIPS codes.
}
