% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\name{get_census_data}
\alias{get_census_data}
\title{Get Census data of all state block groups}
\usage{
get_census_data(us_fips_list, vintage = 2022)
}
\arguments{
\item{us_fips_list}{A vector of state FIPS codes which Census data is to be retrieved.}

\item{vintage}{The vintage year is Census data (default is 2022).}
}
\value{
A dataframe containing Census data all state block groups.
}
\description{
This function retrieves Census data of all state block groups by looping
over the specified list of state FIPS codes.
}
