% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycomb_map_by_acog_districts.R
\name{honeycomb_generate_maps}
\alias{honeycomb_generate_maps}
\title{Generate Hexagon Maps by ACOG District}
\usage{
honeycomb_generate_maps(
  physician_file,
  acog_districts_file,
  trait_map = "all",
  honey_map = "all",
  grid_size = c(0.3, 0.3),
  specific_district = NULL
)
}
\arguments{
\item{physician_file}{A character string specifying the file path to the physician data in RDS format.}

\item{acog_districts_file}{A character string specifying the file path to the ACOG districts data in CSV format.}

\item{trait_map}{A string specifying the trait map (default is "all").}

\item{honey_map}{A string specifying the honey map (default is "all").}

\item{grid_size}{A numeric vector of length 2 specifying the grid size for the hexagon map (default is c(0.3, 0.3)).}

\item{specific_district}{A string or NULL to specify a specific district for generating the map (default is NULL, which processes all districts).}

\item{physician_sf}{A sf object containing physician data with coordinates.}

\item{acog_districts_sf}{A sf object containing the grouped ACOG districts.}
}
\value{
A ggplot object of the generated map for the specified or all districts.

A ggplot object of the generated map for the specified or all districts.
}
\description{
This function generates hexagon maps for ACOG districts.
}
\examples{
\dontrun{
# Generate hexagon maps for a specific district (e.g., District V) using provided data paths
all_map <- honeycomb_generate_maps(
  physician_file = "data/Physicians.rds",
  acog_districts_file = "data/ACOG_Districts.csv",
  trait_map = "all",
  honey_map = "all",
  grid_size = c(0.2, 0.2),
  specific_district = "District V"
)

# Save the generated map
ggsave("all_map.png", all_map, width = 10, height = 6, dpi = 800)
}

}
