% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrr.R
\name{hrr}
\alias{hrr}
\title{Get Hospital Referral Region Shapefile}
\usage{
hrr(remove_HI_AK = TRUE)
}
\arguments{
\item{remove_HI_AK}{Logical, should Hawaii and Alaska be removed? Default is TRUE.}
}
\value{
An sf object containing the hospital referral region data.
}
\description{
This function loads the hospital referral region shapefile and optionally removes Hawaii and Alaska.
}
