% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrr.R
\name{hrr_generate_maps}
\alias{hrr_generate_maps}
\title{Generate Hexagon Maps for Hospital Referral Regions (HRR)}
\usage{
hrr_generate_maps(physician_sf, trait_map = "all", honey_map = "all")
}
\arguments{
\item{physician_sf}{An sf object containing physician data with coordinates.}

\item{trait_map}{A string specifying the trait map (default is "all").}

\item{honey_map}{A string specifying the honey map (default is "all").}
}
\value{
A ggplot object of the generated map.
}
\description{
This function generates hexagon maps for hospital referral regions.
}
