% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create_block_group_overlap.R
\name{map_create_block_group_overlap}
\alias{map_create_block_group_overlap}
\title{Function to create and export a map showing block group overlap with isochrones}
\usage{
map_create_block_group_overlap(
  bg_data,
  isochrones_data,
  output_dir = "figures/"
)
}
\arguments{
\item{bg_data}{A SpatialPolygonsDataFrame representing block group data.}

\item{isochrones_data}{A SpatialPolygonsDataFrame representing isochrone data.}

\item{output_dir}{Directory path for exporting the map files. Default is "figures/".}
}
\value{
None
}
\description{
This function creates a map that displays block groups and their overlap with isochrones.
}
\examples{
\dontrun{
# Create and export the map with the default output directory
map_create_block_group_overlap(block_groups, isochrones_joined_map)

# Create and export the map with a custom output directory
map_create_block_group_overlap(block_groups, isochrones_joined_map, "custom_output/")
}

}
\seealso{
Other mapping: 
\code{\link{create_density_plot}()},
\code{\link{create_individual_isochrone_plots}()},
\code{\link{create_isochrones}()},
\code{\link{create_isochrones_for_dataframe}()},
\code{\link{create_line_plot}()},
\code{\link{create_scatter_plot}()},
\code{\link{map_create_acog_districts_sf}()},
\code{\link{map_create_base}()},
\code{\link{map_create_leaflet_base}()},
\code{\link{map_create_physician_dot}()}
}
\concept{mapping}
