% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinTable.R
\name{MinTable}
\alias{MinTable}
\title{Calculate the Minimum Value(s) and Corresponding Level(s) of a Factor Variable}
\usage{
MinTable(InVec, mult = FALSE)
}
\arguments{
\item{InVec}{Input vector, expected to be a factor variable or convertible to a factor.}

\item{mult}{Logical value indicating whether to return multiple minimum values or just the first one. Default is FALSE.}
}
\value{
If \code{mult} is FALSE, returns the level corresponding to the minimum value of the factor variable.
If \code{mult} is TRUE, returns a character vector containing all the levels with the minimum value.
}
\description{
This function returns the level(s) corresponding to the minimum value(s) of a factor variable.
}
\examples{
vec <- factor(c("A", "B", "A", "C", "B", "B"))
MinTable(vec) # Returns "C"
MinTable(vec, mult = TRUE) # Returns "C"
}
