% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_checks.R
\name{mysterycall_estimate_resources}
\alias{mysterycall_estimate_resources}
\title{Estimate workflow resources}
\usage{
mysterycall_estimate_resources(n_rows)
}
\arguments{
\item{n_rows}{Integer number of input rows used to project runtime and
memory requirements.}
}
\value{
List with runtime and memory estimates
}
\description{
Estimate workflow resources
}
\seealso{
Other utilities: 
\code{\link{mysterycall_download_file}()},
\code{\link{mysterycall_format_pct}()},
\code{\link{mysterycall_remove_constants}()},
\code{\link{mysterycall_remove_near_zero}()},
\code{\link{mysterycall_save_quality_table}()},
\code{\link{mysterycall_assess_data_quality}()},
\code{\link{mysterycall_check_api_response}()},
\code{\link{mysterycall_check_data_completeness}()},
\code{\link{mysterycall_check_dependencies}()},
\code{\link{mysterycall_check_no_data_loss}()},
\code{\link{mysterycall_check_no_limits}()},
\code{\link{mysterycall_export_with_backup}()},
\code{\link{mysterycall_preflight_check}()},
\code{\link{mysterycall_quality_tier}()},
\code{\link{mysterycall_resolve_path}()},
\code{\link{mysterycall_scan_for_limits}()},
\code{\link{mysterycall_standard_labels}()},
\code{\link{mysterycall_standard_palette}()},
\code{\link{mysterycall_use_quiet_logging}()}
}
\concept{utilities}
