% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\name{mysterycall_get_census_data}
\alias{mysterycall_get_census_data}
\title{Get Census data of all state block groups}
\usage{
mysterycall_get_census_data(
  us_fips_list,
  vintage = 2022,
  api_key = Sys.getenv("CENSUS_API_KEY")
)
}
\arguments{
\item{us_fips_list}{A vector of state FIPS codes which Census data is to be retrieved.}

\item{vintage}{The vintage year of Census data (default is 2022).}

\item{api_key}{Census API key to use for requests. Defaults to the \code{CENSUS_API_KEY} environment variable.}
}
\value{
A tibble containing Census block group data for the requested
states. The result includes padded 2020 FIPS columns (\code{statefp},
\code{countyfp}, \code{tractce}, \code{block_group}, \code{geoid}) and a \code{vintage} column
indicating the Census vintage that supplied the estimates.
}
\description{
This function retrieves Census data of all state block groups by looping
over the specified list of state FIPS codes.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
us_fips_list <- c("01", "02")
census_df <- mysterycall_get_census_data(us_fips_list)
\dontshow{\}) # examplesIf}
}
\seealso{
Other census: 
\code{\link{mysterycall_plot_census_age}()},
\code{\link{mysterycall_summarize_census}()}
}
\concept{census}
