% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinician_data.R
\name{mysterycall_get_clinician_data}
\alias{mysterycall_get_clinician_data}
\title{Retrieve Clinician Data}
\usage{
mysterycall_get_clinician_data(input_data)
}
\arguments{
\item{input_data}{Either a data frame containing NPI numbers or a path to a CSV file.}
}
\value{
A tibble with clinician data for the provided NPIs.
}
\description{
This function retrieves clinician data for each valid NPI in the input data frame.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
clinician_df <- mysterycall_get_clinician_data("clinicians.csv")
\dontshow{\}) # examplesIf}
}
\seealso{
Other npi: 
\code{\link{mysterycall_search_and_process_npi}()},
\code{\link{mysterycall_search_taxonomy}()}
}
\concept{npi}
