% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_table.R
\name{mysterycall_max_table}
\alias{mysterycall_max_table}
\title{Calculate the Maximum Value(s) and Corresponding Level(s) of a Factor Variable}
\usage{
mysterycall_max_table(InVec, mult = FALSE)
}
\arguments{
\item{InVec}{Input vector, expected to be a factor variable or convertible to a factor.}

\item{mult}{Logical value indicating whether to return multiple maximum values or just the first one. Default is FALSE.}
}
\value{
Character scalar (\code{mult = FALSE}) or character vector (\code{mult = TRUE})
containing the factor level(s) with the maximum frequency. Returns
\code{character(0)} for a zero-length input.
}
\description{
This function returns the level(s) corresponding to the maximum value(s) of a factor variable.
}
\examples{
vec <- factor(c("A", "B", "A", "C", "B", "B"))
mysterycall_max_table(vec)           # "B"
mysterycall_max_table(vec, mult = TRUE)  # "B"
}
\seealso{
\code{\link[=mysterycall_min_table]{mysterycall_min_table()}}

Other table helpers: 
\code{\link{mysterycall_disparities_table}()},
\code{\link{mysterycall_format_pct}()},
\code{\link{mysterycall_min_table}()},
\code{\link{mysterycall_model_table}()},
\code{\link{mysterycall_table_percentages}()},
\code{\link{mysterycall_table_proportion}()},
\code{\link{print.mysterycall_disparities_table}()}
}
\concept{table helpers}
