% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysterycall_progress_tracker.R
\name{mysterycall_progress_update}
\alias{mysterycall_progress_update}
\alias{mysterycall_progress_update}
\title{Emit a manual progress update}
\usage{
mysterycall_progress_update(tracker, force = FALSE)
}
\arguments{
\item{tracker}{Object created by \code{\link[=mysterycall_progress_tracker]{mysterycall_progress_tracker()}}.}

\item{force}{Logical flag indicating whether the update should be emitted
even if the configured interval has not elapsed.}
}
\description{
Emit a manual progress update
}
\seealso{
Other logging utilities: 
\code{\link{mysterycall_progress_tracker}()},
\code{\link{mysterycall_progress_fail}()},
\code{\link{mysterycall_progress_finish}()},
\code{\link{mysterycall_progress_start}()},
\code{\link{mysterycall_progress_summary}()},
\code{\link{mysterycall_log_cache_hit}()},
\code{\link{mysterycall_log_error}()},
\code{\link{mysterycall_log_info}()},
\code{\link{mysterycall_log_progress}()},
\code{\link{mysterycall_log_save}()},
\code{\link{mysterycall_log_step}()},
\code{\link{mysterycall_log_step_complete}()},
\code{\link{mysterycall_log_success}()},
\code{\link{mysterycall_log_warning}()},
\code{\link{mysterycall_workflow_start}()}
}
\concept{logging utilities}
