% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_constant_vars.R
\name{mysterycall_remove_constants}
\alias{mysterycall_remove_constants}
\title{Remove Constant Variables from a Data Frame}
\usage{
mysterycall_remove_constants(data_frame)
}
\arguments{
\item{data_frame}{A data frame from which constant variables should be removed.}
}
\value{
A data frame with constant variables removed.
}
\description{
This function takes a data frame and returns a new data frame with constant variables removed.
}
\examples{
df <- data.frame(a = 1:3, b = c(5, 5, 5), c = c("x", "y", "z"))
mysterycall_remove_constants(df)

}
\seealso{
Other utilities: 
\code{\link{mysterycall_assess_data_quality}()},
\code{\link{mysterycall_check_api_response}()},
\code{\link{mysterycall_check_data_completeness}()},
\code{\link{mysterycall_check_dependencies}()},
\code{\link{mysterycall_check_no_data_loss}()},
\code{\link{mysterycall_check_no_limits}()},
\code{\link{mysterycall_download_file}()},
\code{\link{mysterycall_estimate_resources}()},
\code{\link{mysterycall_export_with_backup}()},
\code{\link{mysterycall_format_pct}()},
\code{\link{mysterycall_preflight_check}()},
\code{\link{mysterycall_quality_tier}()},
\code{\link{mysterycall_remove_near_zero}()},
\code{\link{mysterycall_resolve_path}()},
\code{\link{mysterycall_save_quality_table}()},
\code{\link{mysterycall_scan_for_limits}()},
\code{\link{mysterycall_standard_labels}()},
\code{\link{mysterycall_standard_palette}()}
}
\concept{utilities}
