% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_near_zero_var.R
\name{mysterycall_remove_near_zero}
\alias{mysterycall_remove_near_zero}
\title{Remove Near-Zero Variance Variables from a Data Frame}
\usage{
mysterycall_remove_near_zero(data_frame, freqCut = 19, uniqueCut = 10)
}
\arguments{
\item{data_frame}{A data frame from which near-zero variance variables should be removed.}

\item{freqCut}{The ratio of the most common value to the second most common value. Defaults to 19.}

\item{uniqueCut}{The percentage of distinct values out of the number of total samples. Defaults to 10.}
}
\value{
A data frame with near-zero variance variables removed.
}
\description{
This function takes a data frame and returns a new data frame with near-zero variance variables removed.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
df <- data.frame(a = 1:20, b = c(rep(1, 19), 2))
mysterycall_remove_near_zero(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other utilities: 
\code{\link{\%>\%}()},
\code{\link{mysterycall_assess_data_quality}()},
\code{\link{mysterycall_check_api_response}()},
\code{\link{mysterycall_check_data_completeness}()},
\code{\link{mysterycall_check_dependencies}()},
\code{\link{mysterycall_check_no_data_loss}()},
\code{\link{mysterycall_check_no_limits}()},
\code{\link{mysterycall_download_file}()},
\code{\link{mysterycall_estimate_resources}()},
\code{\link{mysterycall_export_with_backup}()},
\code{\link{mysterycall_preflight_check}()},
\code{\link{mysterycall_quality_tier}()},
\code{\link{mysterycall_remove_constants}()},
\code{\link{mysterycall_resolve_path}()},
\code{\link{mysterycall_save_quality_table}()},
\code{\link{mysterycall_scan_for_limits}()},
\code{\link{mysterycall_standard_labels}()},
\code{\link{mysterycall_standard_palette}()}
}
\concept{utilities}
