% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_indicators.R
\docType{data}
\name{NCI_CANCER_CENTERS}
\alias{NCI_CANCER_CENTERS}
\title{NCI-Designated Cancer Center Patterns}
\format{
Character vector of NCI cancer center text patterns.
}
\usage{
NCI_CANCER_CENTERS
}
\description{
String patterns for National Cancer Institute designated cancer centers.
Confidence: 0.98.
}
\examples{
NCI_CANCER_CENTERS

}
\seealso{
\code{\link{mysterycall_classify_academic_affiliation}}

Other academic-indicators: 
\code{\link{ACADEMIC_HOSPITAL_PATTERNS}},
\code{\link{ACGME_PROGRAM_INDICATORS}},
\code{\link{COTH_TEACHING_INDICATORS}},
\code{\link{KNOWN_ACADEMIC_INSTITUTIONS}},
\code{\link{MEDICAL_SCHOOL_INDICATORS}},
\code{\link{MEDICARE_GME_INDICATORS}},
\code{\link{NIH_CTSA_HUBS}}
}
\concept{academic-indicators}
\keyword{datasets}
