% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparities_table.R
\name{print.mysterycall_disparities_table}
\alias{print.mysterycall_disparities_table}
\title{Print a mysterycall_disparities_table}
\usage{
\method{print}{mysterycall_disparities_table}(x, ...)
}
\arguments{
\item{x}{A \code{mysterycall_disparities_table} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a mysterycall_disparities_table
}
