% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_and_process_isochrones.R
\name{process_and_save_isochrones}
\alias{process_and_save_isochrones}
\title{Process and Save Isochrones}
\usage{
process_and_save_isochrones(input_file, chunk_size = 25)
}
\arguments{
\item{input_file}{A data frame containing location data with columns "lat" and "long."
The input file should represent geographic coordinates for which
isochrones will be calculated.}

\item{chunk_size}{The number of rows to process in each chunk. Default is 25.}
}
\value{
An sf (simple features) data frame containing isochrone polygons.
}
\description{
This function takes an input file of locations, retrieves isochrones for each location,
and saves them as shapefiles. It processes the data in chunks of 25 rows at a time to
prevent data loss in case of errors.
}
\details{
This function uses the \code{hereR} package to calculate isochrones based on the
provided geographic coordinates. It retrieves isochrones for each location in
the input file, processes the data in chunks to minimize the risk of data loss
in case of errors, and saves the isochrones as shapefiles for further analysis.
}
\examples{
# Load the input file (e.g., from a CSV)
input_file <- read_csv("data/locations.csv")

# Process and save isochrones for the input file (chunk size set to 25)
isochrones_data <- process_and_save_isochrones(input_file, chunk_size = 25)

# Optionally, write the combined isochrones to a shapefile
sf::st_write(isochrones_data, dsn = "data/isochrones/isochrones_all_combined",
             layer = "isochrones", driver = "ESRI Shapefile", quiet = FALSE)

}
