% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_tracker.R
\name{progress_tracker_update}
\alias{progress_tracker_update}
\title{Emit a manual progress update}
\usage{
progress_tracker_update(tracker, force = FALSE)
}
\arguments{
\item{tracker}{Object created by \code{\link[=progress_tracker]{progress_tracker()}}.}

\item{force}{Logical flag indicating whether the update should be emitted
even if the configured interval has not elapsed.}
}
\description{
Emit a manual progress update
}
