% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_constant_vars.R
\name{remove_constant_vars}
\alias{remove_constant_vars}
\title{Remove Constant Variables from a Data Frame}
\usage{
remove_constant_vars(data_frame)
}
\arguments{
\item{data_frame}{A data frame from which constant variables should be removed.}
}
\value{
A data frame with constant variables removed.
}
\description{
This function takes a data frame and returns a new data frame with constant variables removed.
}
\examples{
\dontrun{
new_data <- remove_constant_vars(data_frame)
}

}
