% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_near_zero_var.R
\name{remove_near_zero_var}
\alias{remove_near_zero_var}
\title{Remove Near-Zero Variance Variables from a Data Frame}
\usage{
remove_near_zero_var(data_frame, freqCut = 19, uniqueCut = 10)
}
\arguments{
\item{data_frame}{A data frame from which near-zero variance variables should be removed.}

\item{freqCut}{The ratio of the most common value to the second most common value. Defaults to 19.}

\item{uniqueCut}{The percentage of distinct values out of the number of total samples. Defaults to 10.}
}
\value{
A data frame with near-zero variance variables removed.
}
\description{
This function takes a data frame and returns a new data frame with near-zero variance variables removed.
}
\examples{
\dontrun{
new_data <- remove_near_zero_var(data_frame)
}

}
