% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_quality_check_table.R
\name{save_quality_check_table}
\alias{save_quality_check_table}
\title{Save Quality Check Table}
\usage{
save_quality_check_table(df, filepath)
}
\arguments{
\item{df}{A data frame containing the columns 'npi' and 'name'.}

\item{filepath}{The path where the CSV file should be saved.}
}
\value{
Prints a message to the console indicating that the CSV file has been saved successfully.
}
\description{
This function takes a data frame containing 'npi' and 'name' columns and creates a quality check table.
The table includes the count of observations for each 'npi' and 'name' combination where the count is greater than 2.
The resulting table is saved as a CSV file.
}
