% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitting_dataframe_to_send_to_callers.R
\name{split_and_save}
\alias{split_and_save}
\title{Split data into multiple parts and save each part as separate Excel files.}
\usage{
split_and_save(
  data_or_path,
  output_directory,
  lab_assistant_names,
  seed = 1978,
  complete_file_prefix = "complete_non_split_version_",
  split_file_prefix = "",
  recursive_create = TRUE
)
}
\arguments{
\item{data_or_path}{Either a dataframe containing the input data or a path to the input data file (RDS, CSV, or XLS/XLSX).}

\item{output_directory}{Directory where output Excel files will be saved.}

\item{lab_assistant_names}{Vector of lab assistant names to name the output files.}

\item{seed}{Seed value for randomization (default is 1978).}

\item{complete_file_prefix}{Prefix for the complete output file name (default is "complete_non_split_version_").}

\item{split_file_prefix}{Prefix for each split output file name (default is empty).}

\item{recursive_create}{Logical indicating if directories should be created recursively (default is TRUE).}
}
\description{
Split data into multiple parts and save each part as separate Excel files.
}
\examples{
\dontrun{
library(tyler)
input_data <- readr::read_csv("/path/to/your/input/file.csv")
output_directory <- "/path/to/your/output/directory"
lab_assistant_names <- c("Label1", "Label2", "Label3")
split_and_save(data_or_path = input_data, output_directory, lab_assistant_names)
}
}
