% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_generate_overall.R
\name{table_write_pdf}
\alias{table_write_pdf}
\title{Write an Arsenal table to a PDF file}
\usage{
table_write_pdf(object, filename)
}
\arguments{
\item{object}{An \code{arsenal} table object to write.}

\item{filename}{Path to the output PDF file.}
}
\description{
Utility wrapper around \code{arsenal::write2pdf} to save a table object
as a PDF document.
}
\examples{
\dontrun{
table_write_pdf(overall_summary, "table.pdf")
}
}
