% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy Codes for Obstetricians and Gynecologists}
\format{
A data frame with two columns:
\describe{
\item{NUCC Code}{NUCC (National Uniform Claim Committee) code for healthcare providers.}
\item{Provider Type}{The type of healthcare provider corresponding to the NUCC code.}
}
}
\source{
\url{https://www.nucc.org/images/stories/PDF/taxonomy_23_0.pdf}
}
\usage{
taxonomy
}
\value{
A tibble mapping NUCC taxonomy codes to descriptive provider
types relevant to obstetrics and gynecology research.
}
\description{
This dataset contains taxonomy codes for Obstetricians and Gynecologists among other healthcare providers.
}
\examples{
\dontrun{
# Load the taxonomy dataset
data(taxonomy)

# Explore the dataset
print(taxonomy)
}
}
\seealso{
Other datasets: 
\code{\link{ACOG_Districts}},
\code{\link{acgme}},
\code{\link{cityStateToLatLong}},
\code{\link{fips}}
}
\concept{datasets}
\keyword{datasets}
