\name{test_and_process_isochrones}
\alias{test_and_process_isochrones}
\title{Deprecated isochrone utility}
\usage{
test_and_process_isochrones(input_file, ...)
}
\arguments{
\item{input_file}{Ignored.}

\item{...}{Ignored.}
}
\description{
\strong{Deprecated}. This helper no longer performs any work.
}
\details{
Transition to \code{\link{create_isochrones_for_dataframe}} for supported
behaviour.
}
