% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overall_table.R
\name{tm_write2pdf}
\alias{tm_write2pdf}
\title{Write an Arsenal table to a PDF file}
\usage{
tm_write2pdf(object, filename)
}
\arguments{
\item{object}{An \code{arsenal} table object to write.}

\item{filename}{Path to the output PDF file.}
}
\value{
Invisibly returns the file path.
}
\description{
Utility wrapper around \code{arsenal::write2pdf} to save a table object
as a PDF document.
}
\examples{
\dontrun{
tm_write2pdf(overall_summary, "table.pdf")
}
}
