% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler-package.R
\docType{package}
\name{tyler-package}
\alias{tyler}
\alias{tyler-package}
\title{tyler: Common Functions For Mystery Caller Or Audit Studies Evaluating Patient Access to Care.}
\description{
The tyler package provides functions to perform mystery caller studies. Part of these studies are search and process National Provider Identifier (NPI) numbers. It includes a function to search for NPI numbers based on first and last names and process the search results. The package also offers functionality to generate tables summarizing NPI search and demographic data. The package aims to simplify the process of working with NPI data and generating informative tables for analysis and reporting.
}
\author{
\strong{Maintainer}: Tyler Muffly \email{tyler.muffly@dhha.com}

}
\keyword{internal}
