% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler-package.R
\docType{package}
\name{tyler-package}
\alias{tyler}
\alias{tyler-package}
\title{tyler: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tyler Muffly \email{tyler.muffly@dhha.org} (\href{https://orcid.org/0000-0002-2044-1693}{ORCID})

}
\keyword{internal}
