% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_log_cache_hit}
\alias{tyler_log_cache_hit}
\title{Log cache hit}
\usage{
tyler_log_cache_hit(what, n_items)
}
\arguments{
\item{what}{What was loaded from cache}

\item{n_items}{Number of cached items}
}
\value{
Invisible NULL
}
\description{
Log cache hit
}
