% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_log_error}
\alias{tyler_log_error}
\title{Log error message with context}
\usage{
tyler_log_error(msg, cause = NULL, fix = NULL, indent = TRUE)
}
\arguments{
\item{msg}{Error message}

\item{cause}{Root cause of error}

\item{fix}{Suggested fix}

\item{indent}{Whether to indent (default TRUE)}
}
\value{
Invisible NULL
}
\description{
Log error message with context
}
