% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_library.R
\name{tyler_log_info}
\alias{tyler_log_info}
\title{Emit a timestamped log message when quiet mode is disabled}
\usage{
tyler_log_info(message_text, quiet = getOption("tyler.quiet", FALSE))
}
\arguments{
\item{message_text}{Character string to print.}

\item{quiet}{Optional override for quiet behaviour.}
}
\description{
Emit a timestamped log message when quiet mode is disabled
}
