% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_library.R, R/utils-logging.R
\name{tyler_log_info}
\alias{tyler_log_info}
\title{Emit a timestamped log message when quiet mode is disabled}
\usage{
tyler_log_info(msg, indent = TRUE)

tyler_log_info(msg, indent = TRUE)
}
\arguments{
\item{msg}{Message to log}

\item{indent}{Whether to indent (default TRUE)}

\item{message_text}{Character string to print.}

\item{quiet}{Optional override for quiet behaviour.}
}
\value{
Invisible NULL
}
\description{
Emit a timestamped log message when quiet mode is disabled

Log informational message
}
