% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_log_save}
\alias{tyler_log_save}
\title{Log file save}
\usage{
tyler_log_save(path, n_rows = NULL)
}
\arguments{
\item{path}{File path}

\item{n_rows}{Number of rows (optional)}
}
\value{
Invisible NULL
}
\description{
Log file save
}
