% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_log_success}
\alias{tyler_log_success}
\title{Log success message}
\usage{
tyler_log_success(msg, details = NULL, indent = TRUE)
}
\arguments{
\item{msg}{Message to log}

\item{details}{Optional list of details (name-value pairs)}

\item{indent}{Whether to indent (default TRUE)}
}
\value{
Invisible NULL
}
\description{
Log success message
}
