% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{tyler_log_warning}
\alias{tyler_log_warning}
\title{Log warning message}
\usage{
tyler_log_warning(msg, fix = NULL, indent = TRUE)
}
\arguments{
\item{msg}{Warning message}

\item{fix}{Optional suggested fix}

\item{indent}{Whether to indent (default TRUE)}
}
\value{
Invisible NULL
}
\description{
Log warning message
}
