% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyler_run_workflow_logged.R
\name{tyler_print_dashboard}
\alias{tyler_print_dashboard}
\title{Print a formatted summary dashboard}
\usage{
tyler_print_dashboard(results)
}
\arguments{
\item{results}{List containing workflow results (as returned by
\code{\link[=tyler_run_workflow]{tyler_run_workflow()}}).}
}
\description{
Print a formatted summary dashboard
}
\seealso{
Other workflow: 
\code{\link{tyler_clean_phase1}()},
\code{\link{tyler_clean_phase2}()},
\code{\link{tyler_rename_columns}()},
\code{\link{tyler_run_workflow}()},
\code{\link{tyler_run_workflow_logged}()},
\code{\link{tyler_split_and_save}()}
}
\concept{workflow}
