% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress-bars.R
\name{tyler_progress_fail}
\alias{tyler_progress_fail}
\title{Fail progress bar}
\usage{
tyler_progress_fail(pb, msg = NULL)
}
\arguments{
\item{pb}{Progress bar object from tyler_progress_bar()}

\item{msg}{Error message}
}
\value{
Invisible NULL
}
\description{
Marks progress bar as failed with error message.
}
