% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_library.R
\name{tyler_standard_labels}
\alias{tyler_standard_labels}
\title{Retrieve the standard label dictionary used in \code{tyler}}
\usage{
tyler_standard_labels()
}
\value{
A named list of canonical labels for common columns.
}
\description{
Retrieve the standard label dictionary used in \code{tyler}
}
