% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_remove_invalid_npi.R
\name{validate_and_remove_invalid_npi}
\alias{validate_and_remove_invalid_npi}
\title{Validate and Remove Invalid NPI Numbers}
\usage{
validate_and_remove_invalid_npi(input_data)
}
\arguments{
\item{input_data}{Either a dataframe containing NPI numbers or a path to a CSV file.}
}
\value{
A dataframe containing valid NPI numbers.
}
\description{
This function reads a CSV file containing NPI numbers, validates their
format using the npi package, and removes rows with missing or invalid NPIs.
}
