% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_remove_invalid_npi.R
\name{validate_and_remove_invalid_npi}
\alias{validate_and_remove_invalid_npi}
\title{Validate and Remove Invalid NPI Numbers}
\usage{
validate_and_remove_invalid_npi(input_data)
}
\arguments{
\item{input_csv_path}{Path to the CSV file containing NPI numbers.}
}
\value{
A dataframe containing valid NPI numbers.
}
\description{
This function reads a CSV file containing NPI numbers, validates their
format using the npi package, and removes rows with missing or invalid NPIs.
}
\examples{
input_data <- "~/Dropbox (Personal)/workforce/subspecialists_only.csv"
valid_df <- validate_and_remove_invalid_npi(input_csv_path)

}
